LogicCircuits.jl Documentation

Documentation for LogicCircuits.jl

See the individual modules' API overviews below.

Data Module API

LogicCircuits.Data.AbstractData — Type

Any form of data where features are of type X

LogicCircuits.Data.BatchedXDataset — Type

Batches of unlabeled data with train/validation/test splits

LogicCircuits.Data.BatchedXYDataset — Type

Batches of labeled data with train/validation/test splits

LogicCircuits.Data.Dataset — Type

A dataset with train/validation/test splits

LogicCircuits.Data.LabeledDataset — Type

A labeled dataset with train/validation/test splits

LogicCircuits.Data.PlainXData — Type

Unlabeled data (X-values) with a direct matrix representation

LogicCircuits.Data.UnlabeledDataset — Type

An unlabeled dataset with train/validation/test splits

LogicCircuits.Data.WXData — Type

Unlabeled data (X-values) with weights for each example

LogicCircuits.Data.XBatches — Type

Batches of unlabeled XData

LogicCircuits.Data.XData — Type

Abstract unlabeled data (X-values)

LogicCircuits.Data.XDataset — Type

An single unlabeled dataset with train/validation/test splits

LogicCircuits.Data.XYBatches — Type

Batches of labeled XYData

LogicCircuits.Data.XYData — Type

Labeled data (X-values with Y-labels)

LogicCircuits.Data.XYDataset — Type

An single labeled dataset with train/validation/test splits

LogicCircuits.Data.aggr_weight_type — Method

What type should instance weights aggregate to?

LogicCircuits.Data.feature_data — Method

Get the unlabeled feature portion of the data

LogicCircuits.Data.feature_matrix — Method

Get the data as a feature matrix

LogicCircuits.Data.feature_type — Method

Get the type of the feature matrix elements

LogicCircuits.Data.label_type — Method

Get the type of the label vector elements

LogicCircuits.Data.labels — Method

Get the label vector of the data

LogicCircuits.Data.max_batch_size — Method

Size of the largest batch

LogicCircuits.Data.num_batches — Method

Number of batches in the dataset

LogicCircuits.Data.num_examples — Method

Number of examples in data

LogicCircuits.Data.num_features — Method

Number of features in the data

LogicCircuits.Data.num_labels — Method

Number of labels in the data

LogicCircuits.Data.test — Method

Get the testing fold of the data

LogicCircuits.Data.total_example_weight — Method

Count the number of example in a weighted dataset

LogicCircuits.Data.train — Method

Get the training fold of the data

LogicCircuits.Data.unweighted_data — Method

Get the data as plain unweighted data

LogicCircuits.Data.valid — Method

Get the validation fold of the data

LogicCircuits.Data.vslice — Method

Get slices of data (rows and columns ids) as a view (no memory allocation)

LogicCircuits.Data.weights — Method

Get the weight vector of the data

LogicCircuits.Data.replace_features — Method

Replace the feature matrix in data by a given matrix

LogicCircuits.Data.unslice — Method

Undo batch operation, combine batches into one

Utils Module API

LogicCircuits.Utils — Module

Module with general utilities and missing standard library features that could be useful in any Julia project

Base.typejoin — Method

Get the most specific type parameter possible for an array

LogicCircuits.Utils.children — Method

Get the children of a given inner node

LogicCircuits.Utils.copy_with_eltype — Method

Copy the array while changing the element type

LogicCircuits.Utils.flip_bit — Method

Flip the bit field throughout this circuit (or ensure it is set to given value)

LogicCircuits.Utils.foldup — Method

Compute a function bottom-up on the circuit. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner as a function on the children.

LogicCircuits.Utils.foldup_aggregate — Method

Compute a function bottom-up on the circuit. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner in aggregate as a vector from the children.

LogicCircuits.Utils.grapheltype — Method

Get the type of node contained in this graph

LogicCircuits.Utils.has_children — Method

Does the node have children?

LogicCircuits.Utils.inode_stats — Method

Give count of types and fan-ins of inner nodes in the graph

LogicCircuits.Utils.inodes — Method

Get the list of inner nodes in a given graph

LogicCircuits.Utils.isequal_local — Function

Is one node equal to another locally, ignoring children?

LogicCircuits.Utils.isequal_unordered — Method

Is one unordered tree equal to another?

LogicCircuits.Utils.issomething — Method

Is the argument not nothing?

LogicCircuits.Utils.lca — Method

Find the least common ancestor (assumes the graph has a parent pointer and a list of descendents)

LogicCircuits.Utils.leaf_stats — Method

Give count of types of leaf nodes in the graph

LogicCircuits.Utils.leafnodes — Method

Get the list of leaf nodes in a given graph

LogicCircuits.Utils.left_most_child — Method

Return the leftmost child.

LogicCircuits.Utils.lower_element_type — Method

Specialize the type parameter of an array to be most specific

LogicCircuits.Utils.node2dag — Method

Rebuild a DAG's linear bottom-up order from a new root node

LogicCircuits.Utils.node_stats — Method

Give count of types and fan-ins of all nodes in the graph

LogicCircuits.Utils.num_children — Method

Get the number of children of a given inner node

LogicCircuits.Utils.num_edges — Method

Number of edges in the graph

LogicCircuits.Utils.num_nodes — Method

Number of nodes in the graph

LogicCircuits.Utils.pushrand! — Method

Push element into random position

LogicCircuits.Utils.right_most_child — Method

Return the rightmost child.

LogicCircuits.Utils.tree_num_nodes — Method

Compute the number of nodes in of a tree-unfolding of the DAG.

StatsFuns.logsumexp — Method

Marginalize out dimensions dims from log-probability tensor

LogicCircuits.Utils.Downpass — Type

Compute a function top down on the circuit. f_root is called on the root node, f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed down the circuit and given to f_inner as a value from the parent.

Base.filter — Method

Retrieve list of nodes in circuit matching predicate p

Base.isequal — Method

Is one ordered tree equal to another?

LogicCircuits.Utils.expand_product — Method

generate an expression for the loop-unrolled product of x1 and xs...

LogicCircuits.Utils.expand_product — Method

generate an expression for the loop-unrolled product of xs...

LogicCircuits Module API