Data

LogicCircuits.Data.AbstractData — Type

Any form of data where features are of type X

LogicCircuits.Data.BatchedXDataset — Type

Batches of unlabeled data with train/validation/test splits

LogicCircuits.Data.BatchedXYDataset — Type

Batches of labeled data with train/validation/test splits

LogicCircuits.Data.Dataset — Type

A dataset with train/validation/test splits

LogicCircuits.Data.LabeledDataset — Type

A labeled dataset with train/validation/test splits

LogicCircuits.Data.PlainXData — Type

Unlabeled data (X-values) with a direct matrix representation

LogicCircuits.Data.UnlabeledDataset — Type

An unlabeled dataset with train/validation/test splits

LogicCircuits.Data.WXData — Type

Unlabeled data (X-values) with weights for each example

LogicCircuits.Data.XBatches — Type

Batches of unlabeled XData

LogicCircuits.Data.XData — Type

Abstract unlabeled data (X-values)

LogicCircuits.Data.XDataset — Type

An single unlabeled dataset with train/validation/test splits

LogicCircuits.Data.XYBatches — Type

Batches of labeled XYData

LogicCircuits.Data.XYData — Type

Labeled data (X-values with Y-labels)

LogicCircuits.Data.XYDataset — Type

An single labeled dataset with train/validation/test splits

LogicCircuits.Data.aggr_weight_type — Method

What type should instance weights aggregate to?

LogicCircuits.Data.feature_data — Method

Get the unlabeled feature portion of the data

LogicCircuits.Data.feature_matrix — Method

Get the data as a feature matrix

LogicCircuits.Data.feature_type — Method

Get the type of the feature matrix elements

LogicCircuits.Data.label_type — Method

Get the type of the label vector elements

LogicCircuits.Data.labels — Method

Get the label vector of the data

LogicCircuits.Data.max_batch_size — Method

Size of the largest batch

LogicCircuits.Data.num_batches — Method

Number of batches in the dataset

LogicCircuits.Data.num_examples — Method

Number of examples in data

LogicCircuits.Data.num_features — Method

Number of features in the data

LogicCircuits.Data.num_labels — Method

Number of labels in the data

LogicCircuits.Data.test — Method

Get the testing fold of the data

LogicCircuits.Data.total_example_weight — Method

Count the number of example in a weighted dataset

LogicCircuits.Data.train — Method

Get the training fold of the data

LogicCircuits.Data.unweighted_data — Method

Get the data as plain unweighted data

LogicCircuits.Data.valid — Method

Get the validation fold of the data

LogicCircuits.Data.vslice — Method

Get slices of data (rows and columns ids) as a view (no memory allocation)

LogicCircuits.Data.weights — Method

Get the weight vector of the data

LogicCircuits.Data.replace_features — Method

Replace the feature matrix in data by a given matrix

LogicCircuits.Data.unslice — Method

Undo batch operation, combine batches into one