IO

LogicCircuits.IO.load_logical_circuit — Method

Load a logical circuit from file. Support file formats:

".sdd" for SDD files

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.IO.load_smooth_logical_circuit — Method

Load a smooth logical circuit from file. Support file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.IO.load_struct_smooth_logical_circuit — Method

Load a smooth structured logical circuit from file. Support circuit file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

Supported vtree file formats:

".vtree" for VTree files

LogicCircuits.IO.save — Method

Saves a vtree in the given file path.

LogicCircuits.IO.save_circuit — Method

Save a circuit to file

LogicCircuits.IO.BiasLine — Type

A line representing a bias node in the circuit (an OR with one child)

LogicCircuits.IO.CircuitCommentLine — Type

A string comment line for circuit files

LogicCircuits.IO.CircuitFormatLine — Type

A parsed circuit file format line

LogicCircuits.IO.CircuitFormatLines — Type

A file consisting for circuit formal lines

LogicCircuits.IO.CircuitHeaderLine — Type

A header line for circuit files

LogicCircuits.IO.ConstantLine — Type

A line representing either the true or false logical constants

LogicCircuits.IO.DecisionLine — Type

A line representing a decision node in the circuit (an OR of AND elements)

LogicCircuits.IO.Element — Type

Paired boxes, or elements, are conjunctions in a larger decision node line

LogicCircuits.IO.ID — Type

Circuit and vtree node ids used for IO

LogicCircuits.IO.InnerCircuitLine — Type

A circuit format line with child IDs

LogicCircuits.IO.LeafCircuitLine — Type

A circuit format line without child IDs

LogicCircuits.IO.LiteralLine — Type

A line that represents a logical literal

LogicCircuits.IO.UnweightedLiteralLine — Type

A line representing a single literal without parameters.

LogicCircuits.IO.VtreeAbstractFile — Type

Used to specify file type .vtree or .dot

LogicCircuits.IO.VtreeFormatLine — Type

A line in one vtree file format

LogicCircuits.IO.WeightedLiteralLine — Type

A line representing a weighted single literal (for example a logistic circuit literal).

LogicCircuits.IO.WeightedNamedConstantLine — Type

A weighted constant line for a known variable

LogicCircuits.IO.compile_logical — Method

Compile lines into a unstructured logical circuit

LogicCircuits.IO.compile_logical_m — Method

Compile lines into a unstructured logical circuit, while keeping track of id-to-node mappings

LogicCircuits.IO.compile_smooth_logical — Method

Compile lines into a smooth unstructured logical circuit

LogicCircuits.IO.compile_smooth_logical_m — Method

Compile lines into a smooth unstructured logical circuit, while keeping track of id-to-node mappings

LogicCircuits.IO.compile_smooth_struct_logical — Method

Compile circuit and vtree lines into a structured logical circuit with its vtree

LogicCircuits.IO.compile_smooth_struct_logical_m — Method

Compile circuit and vtree lines into a structured logical circuit with its vtree, while keeping track of id-to-node mappings

LogicCircuits.IO.compile_smooth_struct_logical_m — Method

Compile circuit lines and vtree node mapping into a structured logical circuit, while keeping track of id-to-node mappings

LogicCircuits.IO.decompile — Method

Decompile for sdd circuit, used during saving of circuits to file

LogicCircuits.IO.load_cnf — Method

Load a CNF as a logical circuit from file. Supppor file formats:

".cnf" for CNF files

LogicCircuits.IO.load_dnf — Method

Load a CNF as a logical circuit from file. Supppor file formats:

".cnf" for CNF files

LogicCircuits.IO.save_sdd_file — Method

Save a SDD circuit to file

LogicCircuits.IO.sdd_header — Method

Returns header for SDD file format