LoadSave

LogicCircuits.LoadSave.BiasLine — Type

A line representing a bias node in the circuit (an OR with one child)

LogicCircuits.LoadSave.CircuitCommentLine — Type

A string comment line for circuit files

LogicCircuits.LoadSave.CircuitFormatLine — Type

A parsed circuit file format line

LogicCircuits.LoadSave.CircuitFormatLines — Type

A file consisting for circuit formal lines

LogicCircuits.LoadSave.CircuitHeaderLine — Type

A header line for circuit files

LogicCircuits.LoadSave.ConstantLine — Type

A line representing either the true or false logical constants

LogicCircuits.LoadSave.DecisionLine — Type

A line representing a decision node in the circuit (an OR of AND elements)

LogicCircuits.LoadSave.Element — Type

Paired boxes, or elements, are conjunctions in a larger decision node line

LogicCircuits.LoadSave.ID — Type

Circuit and vtree node ids used for IO

LogicCircuits.LoadSave.InnerCircuitLine — Type

A circuit format line with child IDs

LogicCircuits.LoadSave.LeafCircuitLine — Type

A circuit format line without child IDs

LogicCircuits.LoadSave.LiteralLine — Type

A line that represents a logical literal

LogicCircuits.LoadSave.UnweightedLiteralLine — Type

A line representing a single literal without parameters.

LogicCircuits.LoadSave.VtreeAbstractFile — Type

Used to specify file type .vtree or .dot

LogicCircuits.LoadSave.VtreeFormatLine — Type

A line in the vtree file format

LogicCircuits.LoadSave.WeightedLiteralLine — Type

A line representing a weighted single literal (for example a logistic circuit literal).

LogicCircuits.LoadSave.WeightedNamedConstantLine — Type

A weighted constant line for a known variable

LogicCircuits.LoadSave.cnf_header — Method

Returns header for CNF file format

LogicCircuits.LoadSave.cnf_header_line — Method

Returns header info line for CNF

LogicCircuits.LoadSave.cnf_line — Method

Returns a line of a CNF file for when an AND node has a literal child

LogicCircuits.LoadSave.cnf_line — Method

Returns a line of a CNF file

LogicCircuits.LoadSave.compile_logical — Method

Compile lines into a unstructured logic circuit

LogicCircuits.LoadSave.compile_logical_m — Method

Compile lines into a unstructured logic circuit, while keeping track of id-to-node mappings

LogicCircuits.LoadSave.compile_smooth_logical — Method

Compile lines into a smooth unstructured logic circuit

LogicCircuits.LoadSave.compile_smooth_logical_m — Method

Compile lines into a smooth unstructured logic circuit, while keeping track of id-to-node mappings

LogicCircuits.LoadSave.compile_smooth_struct_logical — Method

Compile circuit and vtree lines into a structured logic circuit with its vtree

LogicCircuits.LoadSave.compile_smooth_struct_logical_m — Method

Compile circuit and vtree lines into a structured logic circuit with its vtree, while keeping track of id-to-node mappings

LogicCircuits.LoadSave.compile_smooth_struct_logical_m — Method

Compile circuit lines and vtree node mapping into a structured logic circuit, while keeping track of id-to-node mappings

LogicCircuits.LoadSave.decompile — Method

Decompile for sdd circuit, used during saving of circuits to file

LogicCircuits.LoadSave.get_level_nodes — Method

Effectively returns the inverse of get_nodes_level, returning a dictionary where keys are the nodes and values are their levels, and the highest level in the circuit.

LogicCircuits.LoadSave.get_nodes_level — Method

Rank nodes in the same layer left to right

LogicCircuits.LoadSave.load_cnf — Method

load_cnf(file::String; dual=false)::PlainLogicCircuit

Load a CNF/DNF as a logic circuit from file. Suppported file formats:

".cnf" for CNF files (dual=False)

".dnf" for DNF files (dual=True)

LogicCircuits.LoadSave.load_dnf — Method

load_dnf(file::String)::PlainLogicCircuit

Load a DNF as a logical circuit from file.

LogicCircuits.LoadSave.load_logic_circuit — Method

load_logic_circuit(file::String)::PlainLogicCircuit

Load a logical circuit from file. Support file formats:

".sdd" for SDD files

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.LoadSave.load_smooth_logic_circuit — Method

load_smooth_logic_circuit(file::String)::PlainLogicCircuit

Load a smooth logic circuit from file. Support file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.LoadSave.load_struct_smooth_logic_circuit — Method

load_struct_smooth_logic_circuit(circuit_file::String, vtree_file::String)::Tuple{StructLogicCircuit,PlainVtree}

Load a smooth structured logic circuit and its vtree from file. Supported circuit file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

Supported vtree file formats:

".vtree" for Vtree files

LogicCircuits.LoadSave.load_vtree — Method

load_vtree(file::String, ::Type{V}=PlainVtree)::V where V<:Vtree

Load a vtree file from file. Currently only supports ".vtree" format.

LogicCircuits.LoadSave.process_mnist — Function

Processes the mnist dataset using the MNIST object from MLDataSets package MLDS_MNIST = the MNIST from MLDataSets labeled = whether to return the lables

LogicCircuits.LoadSave.save_as_cnf — Method

save_as_cnf(name::String, circuit::LogicCircuit)

Save a CNF to file. Circuit should represent a CNF, and file name should end in .cnf

LogicCircuits.LoadSave.save_as_dot — Method

save_as_dot(file::String, circuit::LogicCircuit)

Save logic circuit in 'dot' file format.

LogicCircuits.LoadSave.save_as_dot2tex — Function

save_as_dot2tex(file::String, circuit::LogicCircuit; **other_args)

Save logic circuit as a LaTeX TikZ .tex file using the dot2tex engine.

Argument V is a label map for each vertex; E is a label map for each edge; ⋀_style, ⋁_style, lit_style, ⊤_style, and ⊥_style are styles for each node type; edge_turn is how far down should edges "bend"; node_xsep and node_ysep are x and y-axis separation sizes for terminal nodes.

Note: save_as_tex does not work well on large circuits, and the resulting LaTeX file should serve as a first sketch in (serious) need for adjustments.

LogicCircuits.LoadSave.save_as_sdd — Method

save_as_sdd(name::String, circuit::LogicCircuit, vtree)

Save a SDD circuit to file. File name should end with '.sdd'.

LogicCircuits.LoadSave.save_as_tex — Method

save_as_tex(file::String, circuit::LogicCircuit; **other_args)

Save logic circuit as a LaTeX TikZ .tex file. Argument V is a label map for each vertex, E is a label map for each edge, ⋀_style is the TikZ conjunction node style to use, ⋁_style is the TikZ disjunction node style to use, node_ysep is the y-axis separation distance for nodes, node_xsep for the x-axis, edge_turn is how far down should edges "bend", verbose if set to true prints comments to the LaTeX file for better readability.

Note: save_as_tex does not work well on large circuits, and the resulting LaTeX file should serve as a first sketch in (serious) need for adjustments.

LogicCircuits.LoadSave.save_circuit — Method

save_circuit(name::String, circuit, vtree)

Save a circuit to file. Currently defaults to and only supports saving '.sdd' format.

LogicCircuits.LoadSave.save_vtree — Method

save_vtree(file::AbstractString, vtree::PlainVtree)

Saves a vtree in the given file path based on file format. Supported formats:

".vtree" for Vtree files

".dot" for dot files

LogicCircuits.LoadSave.sdd_header — Method

Returns header for SDD file format

LogicCircuits.LoadSave.twenty_datasets — Method

train, valid, test = twenty_datasets(name)

Load a given dataset from the density estimation datasets. Automatically downloads the files as julia Artifacts. See https://github.com/UCLA-StarAI/Density-Estimation-Datasets for a list of avaialble datasets.