Logical

LogicCircuits.Logical.ConstantGate — Type

A trait denoting constant leaf nodes of any type

LogicCircuits.Logical.ConstantNode — Type

A logical constant leaf node, representing true or false

LogicCircuits.Logical.DecoratorΔ — Type

Any circuit that has an origin represented as a bottom-up linear order of nodes

LogicCircuits.Logical.DecoratorΔNode — Type

A circuit node that has an origin of type O

LogicCircuits.Logical.DownFlowΔ — Method

Construct a downward flow circuit from a given upward flow circuit

LogicCircuits.Logical.FalseNode — Type

Constant False node

LogicCircuits.Logical.FlowΔ — Method

Construct a up and down flow circuit from a given other circuit

LogicCircuits.Logical.GateType — Type

A trait hierarchy denoting types of nodes GateType defines an orthogonal type hierarchy of node types, not circuit types, so we can dispatch on node type regardless of circuit type. See @ref{https://docs.julialang.org/en/v1/manual/methods/#Trait-based-dispatch-1}

LogicCircuits.Logical.GateType — Method

Returns GateType of a node (Literal, Constant, And, Or)

LogicCircuits.Logical.Lit — Type

Literals are represented as 32-bit signed integers. Positive literals are positive integers identical to their variable. Negative literals are their negations. Integer 0 should not be used to represent literals.

LogicCircuits.Logical.LiteralGate — Type

A trait denoting literal leaf nodes of any type

LogicCircuits.Logical.LiteralNode — Type

A logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.Logical.LogicalLeafNode — Type

A logical leaf node

LogicCircuits.Logical.LogicalΔ — Type

A logical circuit represented as a bottom-up linear order of nodes

LogicCircuits.Logical.LogicalΔNode — Type

Root of the logical circuit node hierarchy

LogicCircuits.Logical.PlainVtreeNode — Type

Root of the plain vtree node hierarchy

LogicCircuits.Logical.Sdd — Type

A structured logical circuit represented as a bottom-up linear order of nodes

LogicCircuits.Logical.SddConstantNode — Type

A structured logical constant leaf node, representing true or false. These are the only structured nodes that don't have an associated vtree node (cf. SDD file format)

LogicCircuits.Logical.SddLeafNode — Type

A SDD logical leaf node

LogicCircuits.Logical.SddLiteralNode — Type

A SDD logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.Logical.SddMgrNode — Type

Root of the SDD manager node hierarchy

LogicCircuits.Logical.SddNode — Type

Root of the SDD circuit node hierarchy

LogicCircuits.Logical.Sdd⋀Node — Type

A structured logical conjunction node

LogicCircuits.Logical.Sdd⋁Node — Type

A structured logical disjunction node

LogicCircuits.Logical.StructConstantNode — Type

A structured logical constant leaf node, representing true or false. These are the only structured nodes that don't have an associated vtree node (cf. SDD file format)

LogicCircuits.Logical.StructLiteralNode — Type

A structured logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.Logical.StructLogicalLeafNode — Type

A structured logical leaf node

LogicCircuits.Logical.StructLogicalΔ — Type

A structured logical circuit represented as a bottom-up linear order of nodes

LogicCircuits.Logical.StructLogicalΔNode — Type

Root of the structure logical circuit node hierarchy

LogicCircuits.Logical.Struct⋀Node — Type

A structured logical conjunction node

LogicCircuits.Logical.Struct⋁Node — Type

A structured logical disjunction node

LogicCircuits.Logical.TrimMgrNode — Type

Root of the trimmed SDD manager node hierarchy

LogicCircuits.Logical.TrueNode — Type

Constant True node

LogicCircuits.Logical.UnstLogicalΔ — Type

A unstructured logical circuit represented as a bottom-up linear order of nodes

LogicCircuits.Logical.UnstLogicalΔNode — Type

Root of the unstructured logical circuit node hierarchy

LogicCircuits.Logical.UpFlowΔ — Method

Construct a upward flow circuit from a given other circuit

LogicCircuits.Logical.Var — Type

Variables are represented as 32-bit unsigned integers

LogicCircuits.Logical.VtreeNode — Type

Root of the vtree node hiearchy

LogicCircuits.Logical.Δ — Type

Any circuit represented as a bottom-up linear order of nodes

LogicCircuits.Logical.ΔNode — Type

Root of the circuit node hierarchy

LogicCircuits.Logical.⋀Gate — Type

A trait denoting conjuction nodes of any type

LogicCircuits.Logical.⋀Node — Type

A logical conjunction node (And node)

LogicCircuits.Logical.⋁Gate — Type

A trait denoting disjunction nodes of any type

LogicCircuits.Logical.⋁Node — Type

A logical disjunction node (Or node)

Base.split — Method

Return the circuit after spliting on edge edge and variable var

LogicCircuits.Logical.and_nodes — Method

Get the list of And nodes in a given circuit

LogicCircuits.Logical.balanced_vtree — Method

Construct a balanced vtree

LogicCircuits.Logical.bottom_up_vtree — Method

Construct PlainVtree bottom up, using method specified by combine_method!.

LogicCircuits.Logical.canonicalize — Method

Get the canonical compilation of the given XY Partition

LogicCircuits.Logical.clone — Method

Clone the and node and redirect one of its parents to the new copy

LogicCircuits.Logical.compile — Method

Compile a given variable, literal, or constant

LogicCircuits.Logical.compress — Method

Compress a given XY Partition (merge elements with identical subs)

LogicCircuits.Logical.condition — Method

Return the circuit conditioned on given constrains

LogicCircuits.Logical.conjoin — Method

Conjoin two SDDs

LogicCircuits.Logical.constant — Method

Get the logical constant in a given constant leaf node

LogicCircuits.Logical.constant — Method

Get the logical constant in a given constant leaf node

LogicCircuits.Logical.disjoin — Method

Disjoin two SDDs

LogicCircuits.Logical.false_like — Method

Construct a false node in the hierarchy of node n

LogicCircuits.Logical.forget — Method

Forget variables from the circuit. Warning: this may or may not destroy the determinism property.

LogicCircuits.Logical.fully_factorized_circuit — Method

Generate a fully factorized (logistic regression) circuit over n variables

LogicCircuits.Logical.grand_origin — Method

Get the origin of the origin the given decorator circuit

LogicCircuits.Logical.grand_origin — Method

Get the origin of the origin of the given decorator circuit node

LogicCircuits.Logical.has_unique_constant_nodes — Method

Check whether constant nodes are unique

LogicCircuits.Logical.has_unique_literal_nodes — Method

Check whether literal nodes are unique

LogicCircuits.Logical.is_decomposable — Method

Is the circuit decomposable?

LogicCircuits.Logical.is_false — Method

Is the circuit syntactically equal to false?

LogicCircuits.Logical.is_smooth — Method

Is the circuit smooth?

LogicCircuits.Logical.is_true — Method

Is the circuit syntactically equal to true?

LogicCircuits.Logical.lit2var — Method

Convert a literal its variable, removing the sign of the literal

LogicCircuits.Logical.literal — Method

Get the logical literal in a given literal leaf node

LogicCircuits.Logical.literal — Method

Get the logical literal in a given literal leaf node

LogicCircuits.Logical.model_count — Function

Get the model count of the circuit

LogicCircuits.Logical.negate — Method

Negate an SDD

LogicCircuits.Logical.node_type — Method

Get the node type at the root of the corresponding hierarchy.

LogicCircuits.Logical.num_variables — Method

Number of variables in the circuit

LogicCircuits.Logical.or_nodes — Method

Get the list of or nodes in a given circuit

LogicCircuits.Logical.origin — Method

Get the origin of the given decorator circuit

LogicCircuits.Logical.origin — Method

Get the origin of the given decorator circuit node

LogicCircuits.Logical.origin — Method

Get the first origin of the given decorator circuit node of the given type

LogicCircuits.Logical.origin — Method

Get the first origin of the given decorator circuit of the given node type

LogicCircuits.Logical.path_length — Method

Return the length from vtree inner node n to leaf node which contains var

LogicCircuits.Logical.positive — Method

Get the sign of the literal leaf node

LogicCircuits.Logical.prob_equiv_signature — Method

Get a signature for each node using probabilistic equivalence checking

LogicCircuits.Logical.propagate_constants — Method

Remove all constant leafs from the circuit

LogicCircuits.Logical.sat_prob — Method

Get the probability that a random world satisties the circuit

LogicCircuits.Logical.smooth — Method

Create an equivalent smooth circuit from the given circuit.

LogicCircuits.Logical.top_down_vtree — Method

Construct PlainVtree top town, using method specified by split_method.

LogicCircuits.Logical.true_like — Method

Construct a true node in the hierarchy of node n

LogicCircuits.Logical.unique⋁ — Function

Construct a unique decision gate for the given vtree

LogicCircuits.Logical.var2lit — Method

Convert a variable to the corresponding positive literal

LogicCircuits.Logical.variable — Method

Get the logical variable in a given literal leaf node

LogicCircuits.Logical.variable_scope — Method

Get the variable scope of the circuit node

LogicCircuits.Logical.variable_scope — Method

Get the variable scope of the root of the circuit

LogicCircuits.Logical.variable_scopes — Method

Get the variable scope of each node in the circuit

LogicCircuits.Logical.⋀_nodes — Method

Get the list of conjunction nodes in a given circuit

LogicCircuits.Logical.⋁_nodes — Method

Get the list of disjunction nodes in a given circuit

LogicCircuits.Logical.LogicalInnerNode — Type

A logical inner node

LogicCircuits.Logical.SddInnerNode — Type

A SDD logical inner node

LogicCircuits.Logical.StructLogicalInnerNode — Type

A structured logical inner node

LogicCircuits.Logical.balanced_vtree_root — Method

Construct a balanced vtree root node

LogicCircuits.Logical.canonicalize_compressed — Method

Get the canonical compilation of the given compressed XY Partition

LogicCircuits.Logical.conjoin_cartesian — Method

Conjoin two SDDs when they respect the same vtree node

LogicCircuits.Logical.conjoin_descendent — Method

Conjoin two SDDs when one descends from the other

LogicCircuits.Logical.conjoin_indep — Method

Conjoin two SDDs in separate parts of the vtree

LogicCircuits.Logical.conjoin_like — Method

Conjoin nodes in the same way as the example

LogicCircuits.Logical.conjoin_like — Method

Conjoin nodes in the same way as the example

LogicCircuits.Logical.constant_nodes — Method

Construct a mapping from constants to their canonical node representation

LogicCircuits.Logical.constant_nodes2 — Method

Construct a mapping from constants to their canonical node representation

LogicCircuits.Logical.disjoin_cartesian — Method

Disjoin two SDDs when they respect the same vtree node

LogicCircuits.Logical.disjoin_descendent — Method

Disjoin two SDDs when one descends from the other

LogicCircuits.Logical.disjoin_indep — Method

Disjoin two SDDs in separate parts of the vtree

LogicCircuits.Logical.disjoin_like — Method

Disjoin nodes in the same way as the example

LogicCircuits.Logical.disjoin_like — Method

Disjoin nodes in the same way as the example

LogicCircuits.Logical.disjoin_like — Method

Disjoin nodes in the same way as the example

LogicCircuits.Logical.has_unique_literal_nodes2 — Method

Check whether literal nodes are unique

LogicCircuits.Logical.isvalid — Method

Check vtree validation, variables(parent) = variables(left) + variables(right)

LogicCircuits.Logical.literal_like — Method

Construct a new literal node like the given node's type

LogicCircuits.Logical.literal_like — Method

Construct a new literal node like the given node's type

LogicCircuits.Logical.literal_nodes — Method

Construct a mapping from literals to their canonical node representation

LogicCircuits.Logical.smooth_node — Method

Return a smooth version of the node where the missing_scope variables are added to the scope, using literals from lit_nodes

LogicCircuits.Utils.children — Method

Get the children of a given inner node

LogicCircuits.Utils.foldup — Method

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o through a callback from the children.

LogicCircuits.Utils.foldup_aggregate — Method

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o in an aggregate vector from the children.

LogicCircuits.Utils.isequal_local — Method

Compare whether two vtree nodes are locally equal (enables equals and equals_unordered from Utils)