Logic

LogicCircuits.structfalse — Constant

The unique structured logical false constant

LogicCircuits.structtrue — Constant

The unique structured logical true constant

LogicCircuits.trimfalse — Constant

Canonical trimmed SDD false node

LogicCircuits.trimtrue — Constant

Canonical trimmed SDD true node

LogicCircuits.ApplyCache — Type

Apply cache for the result of conjunctions and disjunctions

LogicCircuits.ConstantGate — Type

A trait denoting constant leaf nodes of any type

LogicCircuits.Element — Type

Represents elements that are not yet compiled into conjunctions

LogicCircuits.GateType — Type

A trait hierarchy denoting types of nodes GateType defines an orthogonal type hierarchy of node types, not circuit types, so we can dispatch on node type regardless of circuit type. See @ref{https://docs.julialang.org/en/v1/manual/methods/#Trait-based-dispatch-1}

LogicCircuits.GateType — Method

Get the gate type trait of the given LogicCircuit

LogicCircuits.InnerGate — Type

A logical gate that is an inner node

LogicCircuits.LeafGate — Type

A logical gate that is a leaf node

LogicCircuits.LiteralGate — Type

A trait denoting literal leaf nodes of any type

LogicCircuits.LogicCircuit — Type

Root of the logic circuit node hierarchy

LogicCircuits.PlainConstantNode — Type

A plain logical constant leaf node, representing true or false

LogicCircuits.PlainFalseNode — Type

Plain constant false node

LogicCircuits.PlainLiteralNode — Type

A plain logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.PlainLogicCircuit — Type

Root of the plain logical circuit node hierarchy

LogicCircuits.PlainLogicInnerNode — Type

A plain logical inner node

LogicCircuits.PlainLogicLeafNode — Type

A plain logical leaf node

LogicCircuits.PlainTrueNode — Type

Plain constant true node

LogicCircuits.PlainVtree — Type

Root of the plain vtree node hierarchy

LogicCircuits.Plain⋀Node — Type

A plain logical conjunction node (And node)

LogicCircuits.Plain⋁Node — Type

A plain logical disjunction node (Or node)

LogicCircuits.Sdd — Type

Root of the SDD circuit node hierarchy

LogicCircuits.SddConstantNode — Type

A SDD logical constant leaf node, representing true or false. These are the only structured nodes that don't have an associated vtree node (cf. SDD file format)

LogicCircuits.SddFalseNode — Type

A SDD logical false constant.

LogicCircuits.SddInnerNode — Type

A SDD logical inner node

LogicCircuits.SddLeafNode — Type

A SDD logical leaf node

LogicCircuits.SddLiteralNode — Type

A SDD logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.SddMgr — Type

Root of the SDD manager node hierarchy

LogicCircuits.SddTrueNode — Type

A SDD logical true constant.

LogicCircuits.Sdd⋀Node — Type

A SDD logical conjunction node

LogicCircuits.Sdd⋁Node — Type

A SDD logical disjunction node

LogicCircuits.StructConstantNode — Type

A structured logical constant leaf node, representing true or false. These are the only structured nodes that don't have an associated vtree node (cf. SDD file format)

LogicCircuits.StructFalseNode — Type

A structured logical false constant. Never construct one, use structfalse to access its unique instance

LogicCircuits.StructLiteralNode — Type

A structured logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.StructLogicCircuit — Type

Root of the structure logical circuit node hierarchy

LogicCircuits.StructLogicInnerNode — Type

A structured logical inner node

LogicCircuits.StructLogicLeafNode — Type

A structured logical leaf node

LogicCircuits.StructTrueNode — Type

A structured logical true constant. Never construct one, use structtrue to access its unique instance

LogicCircuits.Struct⋀Node — Type

A structured logical conjunction node

LogicCircuits.Struct⋁Node — Type

A structured logical disjunction node

LogicCircuits.TrimSddMgr — Type

Root of the trimmed SDD manager node hierarchy

LogicCircuits.TrimSddMgrInnerNode — Type

SDD manager inner vtree node for trimmed SDD nodes

LogicCircuits.TrimSddMgrLeafNode — Type

SDD manager leaf vtree node for trimmed SDD nodes

LogicCircuits.Unique⋁Cache — Type

Unique nodes cache for decision nodes

LogicCircuits.Vtree — Type

Root of the vtree node hiearchy

LogicCircuits.XYPartition — Type

Represent an XY-partition that has not yet been compiled into a disjunction

LogicCircuits.⋀Gate — Type

A trait denoting conjuction nodes of any type

LogicCircuits.⋁Gate — Type

A trait denoting disjunction nodes of any type

LogicCircuits.Utils.children — Method

Get the children of a given inner node

LogicCircuits.Utils.foldup — Method

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o through a callback from the children.

LogicCircuits.Utils.foldup_aggregate — Method

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o in an aggregate vector from the children.

LogicCircuits.Utils.lca — Method

Compute the lowest common ancestor of two vtree nodes Warning: this method uses an imcomplete varsubset check for descends_from and is only correct when v and w are part of the same larger vtree.

LogicCircuits.and_nodes — Method

Get the list of And nodes in a given circuit

LogicCircuits.balanced_vtree — Method

Construct a balanced vtree with the given number of variables

LogicCircuits.balanced_vtree — Method

Construct a balanced vtree with variables ranging from first to last (inclusive)

LogicCircuits.bottom_up_vtree — Method

Construct Vtree bottom up, using method specified by combine_method!.

LogicCircuits.canonical_constants — Method

Construct a mapping from constants to their canonical node representation

LogicCircuits.canonical_literals — Method

Construct a mapping from literals to their canonical node representation

LogicCircuits.canonicalize — Method

Get the canonical compilation of the given XY Partition

LogicCircuits.canonicalize_compressed — Method

Get the canonical compilation of the given compressed XY Partition

LogicCircuits.compile — Function

Create a leaf node in the given hierarchy, compiling a Bool constant or a literal

LogicCircuits.compile_clause — Method

Compile a clause into an SDD

LogicCircuits.compile_cnf — Function

Compile a CNF into an SDD

LogicCircuits.compile_cnf_naive — Function

Compile a CNF into an SDD naively, by conjoining clauses in order of the CNF

LogicCircuits.compile_cnf_tree — Function

Compile a CNF into an SDD, bottom up by distributing clauses over vtree nodes

LogicCircuits.compress — Method

Compress a given XY Partition (merge elements with identical subs)

LogicCircuits.conjoin — Function

Conjoin nodes into a single circuit

LogicCircuits.conjoin — Method

Conjoin two SDDs

LogicCircuits.conjoin_cartesian — Method

Conjoin two SDDs when they respect the same vtree node

LogicCircuits.conjoin_descendent — Method

Conjoin two SDDs when one descends from the other

LogicCircuits.conjoin_indep — Method

Conjoin two SDDs in separate parts of the vtree

LogicCircuits.constant — Function

Get the logical constant in a given constant leaf node

LogicCircuits.constant — Method

Get the logical constant in a given constant leaf node

LogicCircuits.depth — Method

Compute the path length from vtree node n to leaf node which contains var

LogicCircuits.disjoin — Function

Disjoin nodes into a single circuit

LogicCircuits.find_leaf — Method

Find the leaf in the vtree that represents the given variable

LogicCircuits.fully_factorized_circuit — Function

Generate a fully factorized circuit over the given range of variables

LogicCircuits.goes_left — Method

Is the variable v contained in the left branch of m?

LogicCircuits.goes_right — Method

Is the variable v contained in the right branch of m?

LogicCircuits.iscanonical — Method

Does the given circuit have canonical Or gates, as determined by a probabilistic equivalence check?

LogicCircuits.isconstantgate — Method

Is the node a constant gate?

LogicCircuits.isdecomposable — Method

Is the circuit decomposable?

LogicCircuits.isfalse — Method

Is the circuit syntactically equal to false?

LogicCircuits.isinnergate — Method

Is the node an inner gate?

LogicCircuits.isliteralgate — Method

Is the node a literal gate?

LogicCircuits.ispositive — Method

Get the sign of the literal leaf node

LogicCircuits.issmooth — Method

Is the circuit smooth?

LogicCircuits.istrue — Method

Is the circuit syntactically equal to true?

LogicCircuits.is⋀gate — Method

Is the node an And gate?

LogicCircuits.is⋁gate — Method

Is the node an Or gate?

LogicCircuits.literal — Method

Get the logical literal in a given literal leaf node

LogicCircuits.model_count — Function

Get the model count of the circuit

LogicCircuits.negate — Method

Negate an SDD

LogicCircuits.or_nodes — Method

Get the list of or nodes in a given circuit

LogicCircuits.prime — Method

Get the prime, that is, the first conjunct

LogicCircuits.prob_equiv_signature — Method

Get a signature for each node using probabilistic equivalence checking. Note that this implentation may not have any formal guarantees as such.

LogicCircuits.sat_prob — Function

Get the probability that a random world satisties the circuit

LogicCircuits.sdd_num_nodes — Method

Count the number of decision nodes in the SDD

LogicCircuits.sdd_size — Method

Count the number of elements in the decision nodes of the SDD

LogicCircuits.sub — Method

Get the sub, that is, the second and last conjunct

LogicCircuits.top_down_root — Method

Construct Vtree top town, using method specified by split_method.

LogicCircuits.tree_formula_string — Method

Get the formula of a given circuit as a string, expanding the formula into a tree

LogicCircuits.unique⋁ — Function

Construct a unique decision gate for the given vtree

LogicCircuits.variable — Function

Get the variable in a vtree leaf

LogicCircuits.variable — Method

Get the logical variable in a given literal leaf node

LogicCircuits.variables_by_node — Method

Get the variable scope of each node in the circuit

LogicCircuits.varsubset_left — Method

Are the variables in n contained in the left branch of m?

LogicCircuits.varsubset_right — Method

Are the variables in n contained in the right branch of m?

LogicCircuits.vtree — Method

Get the vtree corresponding to the argument

LogicCircuits.⋀_nodes — Method

Get the list of conjunction nodes in a given circuit

LogicCircuits.⋁_nodes — Method

Get the list of disjunction nodes in a given circuit