Utils

LogicCircuits.Utils — Module

Module with general utilities and missing standard library features that could be useful in any Julia project

LogicCircuits.Utils.AbstractBitVector — Constant

Retro-fitted super type of all bit vectors

LogicCircuits.Utils.CuBitVector — Type

Custom CUDA version of BitVector (lacking lots of functionality, just a container for now).

LogicCircuits.Utils.Lit — Type

Literals are represented as 32-bit signed integers. Positive literals are positive integers identical to their variable. Negative literals are their negations. Integer 0 should not be used to represent literals.

LogicCircuits.Utils.Var — Type

Variables are represented as 32-bit unsigned integers

Base.eltype — Method

Find a type that can capture all column values

Base.isdisjoint — Method

Are the given sets disjoint (no shared elements)?

LogExpFunctions.logsumexp — Method

Marginalize out dimensions dims from log-probability tensor

LogicCircuits.Utils.always — Method

An array of 100% probabilities for the given element type

LogicCircuits.Utils.bagging_dataset — Method

Returns an array of DataFrames where each DataFrame is randomly sampled from the original dataset data

LogicCircuits.Utils.batch — Function

Create mini-batches

LogicCircuits.Utils.batch_size — Method

Batch size of the dataset

LogicCircuits.Utils.bits_per_pixel — Method

Normalize the given log-likelihood as bits per pixel in data

LogicCircuits.Utils.chunks — Method

Retrieve chunks of bit vector

LogicCircuits.Utils.eachcol_unweighted — Method

Iterate over columns, excluding the sample weight column

LogicCircuits.Utils.example — Method

Get the ith example

LogicCircuits.Utils.feature_values — Method

Get the ith feature values

LogicCircuits.Utils.fully_factorized_log_likelihood — Method

Computer the per-example log-likelihood of a fully factorized ML model on Bool data

LogicCircuits.Utils.get_bit — Method

Retrieve the jth bit from a BitVector chunk

LogicCircuits.Utils.get_weights — Method

Get the weights from a weighted dataset.

LogicCircuits.Utils.impute — Method

Return a copy of Imputed values of X (potentially statistics from another DataFrame)

For example, to impute using same DataFrame:

impute(X; method=:median)

If you want to use another DataFrame to provide imputation statistics:

impute(test_x, train_x; method=:mean)

Supported methods are :median, :mean, :one, :zero

LogicCircuits.Utils.init_array — Method

An array of undetermined values (fast) for the given element type

LogicCircuits.Utils.isbatched — Method

Is the dataset batched?

LogicCircuits.Utils.isbinarydata — Method

Is the dataset binary?

LogicCircuits.Utils.iscomplete — Method

Is the data complete (no missing values)?

LogicCircuits.Utils.iscomplete_col — Method

Is the data column complete (no missing values)?

LogicCircuits.Utils.isfpdata — Method

Is the dataset consisting of floating point data?

LogicCircuits.Utils.isgpu — Method

Check whether data resides on the GPU

LogicCircuits.Utils.issomething — Method

Is the argument not nothing?

LogicCircuits.Utils.isweighted — Method

Is the dataset weighted?

LogicCircuits.Utils.lit2var — Method

Convert a literal its variable, removing the sign of the literal

LogicCircuits.Utils.ll_per_example — Method

Normalize the given log-likelihood by the number of examples in data

LogicCircuits.Utils.make_missing_mcar — Method

make_missing_mcar(d::DataFrame; keep_prob::Float64=0.8)

Returns a copy of dataframe with making some features missing as MCAR, with keep_prob as probability of keeping each feature.

LogicCircuits.Utils.marginal_prob — Method

Compute the marginal prob of each feature in a binary dataset.

LogicCircuits.Utils.never — Method

An array of 0% probabilities for the given element type

LogicCircuits.Utils.noop — Method

Function that does nothing

LogicCircuits.Utils.num_chunks — Method

For binary complete data, how many UInt64 bit strings are needed to store one feature?

LogicCircuits.Utils.num_examples — Method

num_examples(df::DataFrame)

Number of examples in data

LogicCircuits.Utils.num_features — Method

num_features(df::DataFrame)

Number of features in the data

LogicCircuits.Utils.num_variables — Method

Number of variables in the data structure

LogicCircuits.Utils.order_asc — Method

Order the arguments in a tuple in ascending order

LogicCircuits.Utils.pushrand! — Method

Push element into random position in vectorv

LogicCircuits.Utils.random_sample — Function

Randomly draw samples from the dataset with replacement

LogicCircuits.Utils.same_device — Method

Ensure that x resides on the same device as data

LogicCircuits.Utils.shuffle_examples — Method

shuffle_examples(df::DataFrame)

Shuffle the examples in the data

LogicCircuits.Utils.similar! — Method

Reuse a given array if it has the right type and size, otherwise make a new one

LogicCircuits.Utils.soften — Function

Turn binary data into floating point data close to 0 and 1.

LogicCircuits.Utils.split_sample_weights — Method

Split a weighted dataset into unweighted dataset and its corresponding weights.

LogicCircuits.Utils.subseteq_fast — Method

Replacement for BitSet.⊆ that does not allocate a new BitSet

LogicCircuits.Utils.threshold — Method

Threshold a numeric dataset making it binary

LogicCircuits.Utils.to_cpu — Method

Move data to the CPU

LogicCircuits.Utils.to_gpu — Method

Move data to the GPU

LogicCircuits.Utils.uniform — Method

An array of uniform probabilities

LogicCircuits.Utils.var2lit — Method

Convert a variable to the corresponding positive literal

LogicCircuits.Utils.variables — Function

Get the BitSet of variables in the data structure

LogicCircuits.Utils.weigh_samples — Method

Create a weighted copy of the data set