LoadSave

LogicCircuits.LoadSave.BiasLine — Type

A line representing a bias node in the circuit (an OR with one child)

LogicCircuits.LoadSave.CircuitCommentLine — Type

A string comment line for circuit files

LogicCircuits.LoadSave.CircuitFormatLine — Type

A parsed circuit file format line

LogicCircuits.LoadSave.CircuitFormatLines — Type

A file consisting for circuit formal lines

LogicCircuits.LoadSave.CircuitHeaderLine — Type

A header line for circuit files

LogicCircuits.LoadSave.ConstantLine — Type

A line representing either the true or false logical constants

LogicCircuits.LoadSave.DecisionLine — Type

A line representing a decision node in the circuit (an OR of AND elements)

LogicCircuits.LoadSave.Element — Type

Paired boxes, or elements, are conjunctions in a larger decision node line

LogicCircuits.LoadSave.ID — Type

Circuit and vtree node ids used for IO

LogicCircuits.LoadSave.InnerCircuitLine — Type

A circuit format line with child IDs

LogicCircuits.LoadSave.LeafCircuitLine — Type

A circuit format line without child IDs

LogicCircuits.LoadSave.LiteralLine — Type

A line that represents a logical literal

LogicCircuits.LoadSave.UnweightedLiteralLine — Type

A line representing a single literal without parameters.

LogicCircuits.LoadSave.VtreeAbstractFile — Type

Used to specify file type .vtree or .dot

LogicCircuits.LoadSave.VtreeFormatLine — Type

A line in the vtree file format

LogicCircuits.LoadSave.WeightedLiteralLine — Type

A line representing a weighted single literal (for example a logistic circuit literal).

LogicCircuits.LoadSave.WeightedNamedConstantLine — Type

A weighted constant line for a known variable

LogicCircuits.LoadSave.compile_logical — Method

Compile lines into a unstructured logical circuit

LogicCircuits.LoadSave.compile_logical_m — Method

Compile lines into a unstructured logical circuit, while keeping track of id-to-node mappings

LogicCircuits.LoadSave.compile_smooth_logical — Method

Compile lines into a smooth unstructured logical circuit

LogicCircuits.LoadSave.compile_smooth_logical_m — Method

Compile lines into a smooth unstructured logical circuit, while keeping track of id-to-node mappings

LogicCircuits.LoadSave.compile_smooth_struct_logical — Method

Compile circuit and vtree lines into a structured logical circuit with its vtree

LogicCircuits.LoadSave.compile_smooth_struct_logical_m — Method

Compile circuit and vtree lines into a structured logical circuit with its vtree, while keeping track of id-to-node mappings

LogicCircuits.LoadSave.compile_smooth_struct_logical_m — Method

Compile circuit lines and vtree node mapping into a structured logical circuit, while keeping track of id-to-node mappings

LogicCircuits.LoadSave.decompile — Method

Decompile for sdd circuit, used during saving of circuits to file

LogicCircuits.LoadSave.get_level_nodes — Method

Effectively returns the inverse of get_nodes_level, returning a dictionary where keys are the nodes and values are their levels, and the highest level in the circuit.

LogicCircuits.LoadSave.get_nodes_level — Method

Rank nodes in the same layer left to right

LogicCircuits.LoadSave.load_cnf — Method

Load a CNF/DNF as a logical circuit from file. Suppported file formats:

".cnf" for CNF files

".dnf" for DNF files

LogicCircuits.LoadSave.load_logic_circuit — Method

Load a logical circuit from file. Support file formats:

".sdd" for SDD files

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.LoadSave.load_smooth_logic_circuit — Method

Load a smooth logical circuit from file. Support file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.LoadSave.load_struct_smooth_logic_circuit — Method

Load a smooth structured logical circuit from file. Support circuit file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

Supported vtree file formats:

".vtree" for Vtree files

LogicCircuits.LoadSave.save_as_dot — Method

Save logic circuit to .dot file

LogicCircuits.LoadSave.save_as_dot2tex — Function

Save logic circuit as a LaTeX TikZ .tex file using the dot2tex engine.

Argument V is a label map for each vertex; E is a label map for each edge; ⋀_style, ⋁_style, lit_style, ⊤_style, and ⊥_style are styles for each node type; edge_turn is how far down should edges "bend"; node_xsep and node_ysep are x and y-axis separation sizes for terminal nodes.

Note: save_as_tex does not work well on large circuits, and the resulting LaTeX file should serve as a first sketch in (serious) need for adjustments.

LogicCircuits.LoadSave.save_as_sdd — Method

Save a SDD circuit to file

LogicCircuits.LoadSave.save_as_tex — Function

Save logic circuit as a LaTeX TikZ .tex file. Argument V is a label map for each vertex, E is a label map for each edge, ⋀_style is the TikZ conjunction node style to use, ⋁_style is the TikZ disjunction node style to use, node_ysep is the y-axis separation distance for nodes, node_xsep for the x-axis, edge_turn is how far down should edges "bend", verbose if set to true prints comments to the LaTeX file for better readability.

Note: save_as_tex does not work well on large circuits, and the resulting LaTeX file should serve as a first sketch in (serious) need for adjustments.

LogicCircuits.LoadSave.save_circuit — Method

Save a circuit to file

LogicCircuits.LoadSave.save_vtree — Method

Saves a vtree in the given file path.

LogicCircuits.LoadSave.sdd_header — Method

Returns header for SDD file format