LogicCircuits

LogicCircuits.false_sdd — Constant

Canonical false Sdd node

LogicCircuits.structfalse — Constant

The unique splain tructured logical false constant

LogicCircuits.structtrue — Constant

The unique plain structured logical true constant

LogicCircuits.true_sdd — Constant

Canonical true Sdd node

LogicCircuits.ApplyArgs — Type

Representation of the arguments of an Apply call

LogicCircuits.ApplyCache — Type

Apply cache for the result of conjunctions and disjunctions

LogicCircuits.Bdd — Type

A Binary Decision Diagram.

index: the vertex variable (-1 if terminal vertex

low: low child vertex of BDD (undef if terminal vertex)

high: high child vertex of BDD (undef if terminal vertex)

value: terminal boolean value

id: unique identifier

LogicCircuits.BitCircuit — Type

A bit circuit is a low-level representation of a logical circuit structure.

They are a "flat" representation of a circuit, essentially a bit string, that can be processed by lower level code (e.g., GPU kernels)

The wiring of the circuit is captured by two matrices: nodes and elements.

Nodes are either leafs or decision (disjunction) nodes in the circuit.

Elements are conjunction nodes in the circuit.

In addition, there is a vector of layers, where each layer is a list of node ids. Layer 1 is the leaf/input layer. Layer end is the circuit root.

And there is a vector of parents, pointing to element id parents of decision nodes.

Nodes are represented as a 4xN matrix where

nodes[1,:] is the first element id belonging to this decision

nodes[2,:] is the last element id belonging to this decision

nodes[3,:] is the first parent index belonging to this decision

nodes[4,:] is the last parent index belonging to this decision

Elements belonging to node i are elements[:, nodes[1,i]:nodes[2,i]] Parents belonging to node i are parents[nodes[3,i]:nodes[4,i]]

Elements are represented by a 3xE matrix, where

elements[1,:] is the decision node id (parents of the element),

elements[2,:] is the prime node id (child of the element)

elements[3,:] is the sub node id (child of the element)

LogicCircuits.BitCircuit — Method

construct a new BitCircuit accomodating the given number of features

LogicCircuits.ConstantGate — Type

A trait denoting constant leaf nodes of any type

LogicCircuits.Element — Type

Represents elements that are not yet compiled into conjunctions

LogicCircuits.GateType — Type

A trait hierarchy denoting types of nodes GateType defines an orthogonal type hierarchy of node types, not circuit types, so we can dispatch on node type regardless of circuit type. See @ref{https://docs.julialang.org/en/v1/manual/methods/#Trait-based-dispatch-1}

LogicCircuits.GateType — Method

Get the gate type trait of the given LogicCircuit

LogicCircuits.InnerGate — Type

A logical gate that is an inner node

LogicCircuits.LeafGate — Type

A logical gate that is a leaf node

LogicCircuits.LiteralGate — Type

A trait denoting literal leaf nodes of any type

LogicCircuits.LogicCircuit — Type

Root of the logic circuit node hierarchy

LogicCircuits.NodeId — Type

Integer identifier for a circuit node

LogicCircuits.NodeIds — Type

The BitCircuit ids associated with a node

LogicCircuits.PlainConstantNode — Type

A plain logical constant leaf node, representing true or false

LogicCircuits.PlainLiteralNode — Type

A plain logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.PlainLogicCircuit — Type

Root of the plain logic circuit node hierarchy

LogicCircuits.PlainLogicInnerNode — Type

A plain logical inner node

LogicCircuits.PlainLogicLeafNode — Type

A plain logical leaf node

LogicCircuits.PlainStructConstantNode — Type

A plain structured logical constant leaf node, representing true or false. These are the only structured nodes that don't have an associated vtree node (cf. SDD file format)

LogicCircuits.PlainStructLiteralNode — Type

A plain structured logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.PlainStructLogicCircuit — Type

Root of the plain structure logic circuit node hierarchy

LogicCircuits.PlainStructLogicInnerNode — Type

A plain structured logical inner node

LogicCircuits.PlainStructLogicLeafNode — Type

A plain structured logical leaf node

LogicCircuits.PlainStruct⋀Node — Type

A plain structured logical conjunction node

LogicCircuits.PlainStruct⋁Node — Type

A plain structured logical disjunction node

LogicCircuits.PlainVtree — Type

Root of the plain vtree node hierarchy

LogicCircuits.Plain⋀Node — Type

A plain logical conjunction node (And node)

LogicCircuits.Plain⋁Node — Type

A plain logical disjunction node (Or node)

LogicCircuits.Sdd — Type

Root of the trimmed Sdd circuit node hierarchy

LogicCircuits.SddConstantNode — Type

A SDD logical constant leaf node, representing true or false. These are the only structured nodes that don't have an associated vtree node (cf. SDD file format)

LogicCircuits.SddInnerNode — Type

A SDD logical inner node

LogicCircuits.SddLeafNode — Type

A SDD logical leaf node

LogicCircuits.SddLiteralNode — Type

A SDD logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.SddMgr — Type

Root of the SDD manager node hierarchy

LogicCircuits.SddMgrInnerNode — Type

SDD manager inner vtree node for trimmed SDD nodes

LogicCircuits.SddMgrLeafNode — Type

SDD manager leaf vtree node for trimmed SDD nodes

LogicCircuits.Sdd⋀Node — Type

A SDD logical conjunction node

LogicCircuits.Sdd⋁Node — Type

A SDD logical disjunction node

LogicCircuits.StructLogicCircuit — Type

Root of the structure logic circuit node hierarchy

LogicCircuits.Unique⋁Cache — Type

Unique nodes cache for decision nodes

LogicCircuits.Vtree — Type

Root of the vtree node hiearchy

LogicCircuits.XYPartition — Type

Represent an XY-partition that has not yet been compiled into a disjunction

LogicCircuits.⋀Gate — Type

A trait denoting conjuction nodes of any type

LogicCircuits.⋁Gate — Type

A trait denoting disjunction nodes of any type

Base.:!= — Method

Returns whether the two given boolean functions are not equivalent.

Base.:== — Method

Returns whether the two given boolean functions are equivalent.

Base.:| — Method

Returns a new reduced Bdd restricted to instantiation X.

Base.:⊻ — Method

Returns a xor of the given boolean functions.

Base.deepcopy — Function

deepcopy(n::Node, depth::Int64)

Recursively create a copy circuit rooted at n to a certain depth depth

Base.hash — Method

Returns a unique hash for the whole BDD.

Base.merge — Method

merge(root::Node, or1::Node, or2::Node)

Merge two circuits.

Base.sign — Method

Returns 0 if x is not a literal; else returns the literal's sign.

Base.size — Method

Returns the number of nodes in the BDD graph.

Base.split — Method

split(root::Node, (or, and)::Tuple{Node, Node}, var::Var; depth=0, sanity_check=true)

Return the circuit after spliting on edge edge and variable var

Base.string — Method

Return string representation of Bdd α.

Base.xor — Method

Exclusive logical disjunction (XOR)

LogicCircuits.:¬ — Method

Negates this boolean function.

LogicCircuits.:¬ — Method

Logical negation

LogicCircuits.:⇐ — Method

Material logical implication (reverse)

LogicCircuits.:⇒ — Method

Material logical implication

LogicCircuits.:⇔ — Method

Bidirectional logical implication

LogicCircuits.:∧ — Method

Returns a conjunction over the given boolean functions.

LogicCircuits.:∧ — Method

Logical conjunction

LogicCircuits.:∨ — Method

Returns a disjunction over the given boolean functions.

LogicCircuits.:∨ — Method

Logical disjunction

LogicCircuits.LoadSave.load_cnf — Method

Loads a CNF as a BDD. Use load instead.

LogicCircuits.LoadSave.load_dnf — Method

Loads a CNF as a BDD. Use load instead.

LogicCircuits.Utils.children — Method

Get the children of a given inner node

LogicCircuits.Utils.depth — Method

Compute the path length from vtree node n to leaf node for variable var

LogicCircuits.Utils.find_inode — Method

Find an inner vtree node that has left in its left subtree and right in its right subtree. Supports nothing as a catch-all for any node

LogicCircuits.Utils.find_leaf — Method

Find the leaf in the vtree that represents the given variable

LogicCircuits.Utils.foldup — Method

foldup(node::LogicCircuit, 
    f_con::Function, 
    f_lit::Function, 
    f_a::Function, 
    f_o::Function)::T where {T}

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o through a callback from the children.

LogicCircuits.Utils.foldup_aggregate — Method

foldup_aggregate(node::LogicCircuit, 
    f_con::Function, 
    f_lit::Function, 
    f_a::Function, 
    f_o::Function, 
    ::Type{T})::T where T

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o in an aggregate vector from the children.

LogicCircuits.Utils.lca — Method

Compute the lowest common ancestor of two vtree nodes Warning: this method uses an incomplete varsubset check for descends_from and is only correct when v and w are part of the same larger vtree.

LogicCircuits.Utils.num_nodes — Method

How many nodes are indexed by a given bit circuit?

LogicCircuits.Utils.variables — Function

variables(root::LogicCircuit)::BitSet

Get a bitset of variables mentioned in the circuit root.

LogicCircuits.all_valuations — Method

Computes all possible valuations of scope V and returns as a BitMatrix. Up to 64 variables.

LogicCircuits.and — Method

Returns a conjunction over the given boolean functions.

LogicCircuits.and_nodes — Method

Get the list of And nodes in a given circuit

LogicCircuits.apply — Method

Returns a Bdd canonical representation of α ⊕ β, where ⊕ is some binary operator.

LogicCircuits.apply_step — Method

Recursively computes α ⊕ β. If the result was already computed as an intermediate result, return the cached result in T.

LogicCircuits.atleast! — Method

Constructs a BDD mapping to true if at least n literals in L are in the input; otherwise false.

LogicCircuits.atleast — Method

Constructs a BDD mapping to true if at least n literals in L are in the input; otherwise false.

LogicCircuits.atmost! — Method

Constructs a BDD mapping to true if at most n literals in L are in the input; otherwise false.

LogicCircuits.atmost — Method

Constructs a BDD mapping to true if at most n literals in L are in the input; otherwise false.

LogicCircuits.balance_threads — Method

assign threads to examples and decision nodes so that everything is a power of 2

LogicCircuits.bdd_var — Method

Returns a Bdd representing a single variable. If negative, negate variable.

LogicCircuits.canonical_constants — Method

Construct a mapping from constants to their canonical node representation

LogicCircuits.canonical_literals — Method

Construct a mapping from literals to their canonical node representation

LogicCircuits.canonicalize — Method

Get the canonical compilation of the given XY Partition

LogicCircuits.canonicalize_compressed — Method

Get the canonical compilation of the given compressed XY Partition

LogicCircuits.clone — Method

Clone the or node and redirect one of its parents to the new copy

LogicCircuits.compile — Function

Create new circuit nodes in the given context.

LogicCircuits.compile — Method

Compile a given variable, literal, or constant

LogicCircuits.compress — Method

Compress a given XY Partition (merge elements with identical subs)

LogicCircuits.conjoin — Function

Conjoin nodes into a single circuit

LogicCircuits.conjoin — Method

conjoin(root::Node, lit::Lit; callback::Function)::Node

Return the circuit conjoined with th given literal constrains callback is called after modifying conjunction node

LogicCircuits.conjoin — Method

Conjoin two SDDs

LogicCircuits.conjoin_cartesian — Method

Conjoin two SDDs when they respect the same vtree node

LogicCircuits.conjoin_descendent — Method

Conjoin two SDDs when one descends from the other

LogicCircuits.conjoin_indep — Method

Conjoin two SDDs in separate parts of the vtree

LogicCircuits.conjunctions — Method

Computes all possible valuations of scope V as conjunctions.

LogicCircuits.constant — Method

Get the logical constant in a given constant leaf node

LogicCircuits.construct_new_sdd_literal — Method

Helper function to construct new SDD literal objects in SddMgrLeafNode constructor

LogicCircuits.convals — Method

Computes all possible valuations of scope V as both conjunctions and instantiation values.

LogicCircuits.culledfreqs — Method

Returns an approximation (does not account for some repeated nodes) of how many times each variable is mentioned in α.

LogicCircuits.disjoin — Function

Disjoin nodes into a single circuit

LogicCircuits.eliminate — Method

Eliminate a variable through disjunction. Equivalent to the expression (ϕ|x ∨ ϕ|¬x).

LogicCircuits.entails — Method

Decide whether one sentence logically entails another

LogicCircuits.equivalent — Method

Decide whether two sentences are logically equivalent

LogicCircuits.exactly! — Method

Constructs a BDD mapping to true if exactly n literals in L are in the input; otherwise false.

LogicCircuits.exactly — Method

Constructs a BDD mapping to true if exactly n literals in L are in the input; otherwise false.

LogicCircuits.fill_missing_vtree — Method

Construct a smoothed node from startvtr, when you get to endvtr insert the original node

LogicCircuits.forget — Method

Returns the resulting BDD after applying the forget operation. Equivalent to $\phi|_x \vee \phi|_{\neg x}$.

LogicCircuits.forget — Method

forget(root::Node, is_forgotten::Function)::Node

Forget variables from the circuit. Warning: this may or may not destroy the determinism property.

LogicCircuits.from_npbc — Method

Translates a cardinality constraint in normal pseudo-boolean constraint form into a BDD.

Since cardinality constraints correspond to having coefficients set to one, we ignore the C's.

Argument L corresponds to the vector of literals to be chosen from; b is how many literals in L are selected.

See Eén and Sörensson 2006.

LogicCircuits.fully_factorized_circuit — Function

Generate a fully factorized circuit over the given range of variables

LogicCircuits.global_scope — Method

Get the variable scope of the root of this vtree

LogicCircuits.has_single_child — Method

Does the bitcircuit node have a single child?

LogicCircuits.implied_literals — Function

implied_literals(root::LogicCircuit)::Union{BitSet, Nothing}

Compute at each node literals that are implied by the formula. nothing at a node means all literals are implied (i.e. the node's formula is false)

This algorithm is sound but not complete - all literals returned are correct, but some true implied literals may be missing.

LogicCircuits.infer_vtree — Function

infer_vtree(root::LogicCircuit)::Vtree

Infer circuits vtree if the circuit is struct-decomposable it. Otherwise return nothing.

LogicCircuits.init_satisfies — Method

Initialize values from the data (data frames)

LogicCircuits.is_atom — Method

Returns whether the given Bdd node is an atomic formula (i.e. a variable, ⊥, ⊤, or literal).

LogicCircuits.is_lit — Method

Returns whether the given Bdd node represents a literal.

LogicCircuits.is_term — Method

Returns whether this Bdd node is terminal.

LogicCircuits.is_var — Method

Returns whether the given Bdd node represents a variable.

LogicCircuits.is_⊤ — Method

Returns whether the given Bdd node represents a ⊤.

LogicCircuits.is_⊥ — Method

Returns whether the given Bdd node represents a ⊥.

LogicCircuits.iscanonical — Method

iscanonical(circuit::LogicCircuit, k::Int; verbose = false)

Does the given circuit have canonical Or gates, as determined by a probabilistic equivalence check?

LogicCircuits.isconstantgate — Method

Is the node a constant gate?

LogicCircuits.isdecomposable — Function

isdecomposable(root::LogicCircuit)::Bool

Is the circuit decomposable?

LogicCircuits.isdeterministic — Function

isdeterministic(root::LogicCircuit)::Bool

Is the circuit determinstic? Note: this function is generally intractable for large circuits.

LogicCircuits.isfalse — Method

Is the circuit syntactically equal to false?

LogicCircuits.isinnergate — Method

Is the node an inner gate?

LogicCircuits.isleafgate — Method

Is the node a leaf gate?

LogicCircuits.isliteralgate — Method

Is the node a literal gate?

LogicCircuits.ispositive — Method

Get the sign of the literal leaf node

LogicCircuits.issatisfiable — Method

issatisfiable(root::LogicCircuit)::Bool

Determine whether the logical circuit is satisfiable (has a satisfying assignment). Requires decomposability of the circuit.

LogicCircuits.issmooth — Function

issmooth(root::LogicCircuit)::Bool

Is the circuit smooth?

LogicCircuits.isstruct_decomposable — Method

isstruct_decomposable(root::LogicCircuit)::Bool

Is the circuit structured-decomposable?

LogicCircuits.istautology — Method

istautology(root::LogicCircuit)::Bool

Determine whether the logical circuit is a tautology (every assignment satisfies it; the sentence is valid). Requires decomposability and determinism of the circuit.

LogicCircuits.istrue — Method

Is the circuit syntactically equal to true?

LogicCircuits.is⋀gate — Method

Is the node an And gate?

LogicCircuits.is⋁gate — Method

Is the node an Or gate?

LogicCircuits.lit_val — Method

Returns whether a variable x appears as a positive literal in α, given that α is a conjunction of literals.

LogicCircuits.lit_vec — Method

Assumes ϕ is a full conjunction of literals. Returns ϕ as a zero-one vector and its scope.

LogicCircuits.literal — Method

Get the logical literal in a given literal leaf node

LogicCircuits.literal_nodes — Method

Get the list of literal nodes in a given circuit

LogicCircuits.literals — Method

Get a sequence of positive and negative literals

LogicCircuits.load — Method

Loads a BDD from given file.

Supported file formats:

CNF (.cnf);

DNF (.dnf);

BDD (.bdd).

To load as any of these file formats, simply set the filename with the desired extension.

Keyword arguments are passed down to the open function.

LogicCircuits.load_bdd — Method

Loads a BDD from a file. Use load instead.

LogicCircuits.map_parents — Method

Computes a mapping of the parents of each node.

LogicCircuits.marginalize — Method

Marginalize a variable through some binary operation.

LogicCircuits.max_variable — Function

Get the variable in the circuit with the largest index

LogicCircuits.mentions — Method

Returns whether the formula (i.e. BDD) mentions a variable.

LogicCircuits.mgr — Method

Get the manager of a Sdd node, which is its SddMgr vtree

LogicCircuits.model_count — Function

model_count(root::LogicCircuit, num_vars_in_scope::Int = num_variables(root))::BigInt

Get the model count of the circuit. The num_vars_in_scope is set to number of variables in the circuit, but sometimes need to set different values, for example, if not every variable is mentioned in the circuit.

LogicCircuits.model_var_prob — Method

Compute the probability of each variable for a random satisfying assignment of the logical circuit

LogicCircuits.neg_literals — Method

Get a sequence of negative literals

LogicCircuits.negate — Method

Negate an SDD

LogicCircuits.normal_form — Method

Runs a BFS on the mapping of parents, starting from either a ⊤ (true) or ⊥ (false) in order to find the corresponding CNF or DNF encoding.

LogicCircuits.num_elements — Method

Number of elements (conjunctions) in layer or bit circuit

LogicCircuits.or — Method

Returns a disjunction over the given boolean functions.

LogicCircuits.or_nodes — Method

Get the list of or nodes in a given circuit

LogicCircuits.pos_literals — Method

Get a sequence of positive literals

LogicCircuits.postorder — Method

Returns a Vector{Bdd} containing all nodes in α in post-order.

LogicCircuits.prime — Method

Get the prime, that is, the first conjunct

LogicCircuits.print_conjunction — Method

Pretty print a conjunction of literals BDD.

LogicCircuits.print_nf — Method

Pretty print BDD as a normal form (CNF or DNF).

Caution: may exponentially explode.

Alternatively, pretty prints using the given glyphs (default ∧, ∨ and ¬).

ϕ = (1 ∧ ¬2) ∨ (2 ∧ 3)
print_nf(α; out = false)

ϕ = (1 ∧ ¬2) ∨ (2 ∧ 3)
print_nf(α; out = false, which = "dnf", glyphs = ['+', '*', '-'])

LogicCircuits.prob_equiv_signature — Function

prob_equiv_signature(circuit::LogicCircuit, k::Int)::Dict{Union{Var,Node},Signature}

Get a signature for each node using probabilistic equivalence checking. Note that this implentation may not have any formal guarantees as such.

LogicCircuits.propagate_constants — Method

propagate_constants(root::Node)

Remove all constant leafs from the circuit

LogicCircuits.random_split — Method

Randomly picking egde and variable from candidates

LogicCircuits.reduce! — Method

Reduce a Bdd rooted at α inplace, removing duplicate nodes and redundant sub-trees. Returns canonical representation of α.

LogicCircuits.replace_node — Method

Replace node old with node new in circuit root

LogicCircuits.respects_vtree — Method

Does the circuit respect the given vtree? This function allows for constants in conjunctions, but only when a vtree node can be found where the left and right conjunct can be assigned to the left and right vtree.

LogicCircuits.restrict — Method

Returns a new reduced Bdd restricted to instantiation X.

LogicCircuits.restrict_step — Method

Returns a new Bdd restricted to instantiation X.

LogicCircuits.restrict_step — Method

Returns a new Bdd restricted to instantiation X.

LogicCircuits.sat_prob — Function

sat_prob(root::LogicCircuit; varprob::Function)::Rational{BigInt}

Get the probability that a random world satisties the circuit. Probability of each variable is given by varprob Function which defauls to 1/2 for every variable.

LogicCircuits.satisfies — Method

Evaluate satisfaction of the circuit bottom-up for a given input

LogicCircuits.satisfies_all — Function

Evaluate the circuit bottom-up for a given input and return the value of all nodes

LogicCircuits.satisfies_flows — Function

Compute the value and flow of each node

LogicCircuits.satisfies_flows_down — Function

When values of nodes have already been computed, do a downward pass computing the flows at each node

LogicCircuits.satisfies_flows_down_layers — Method

Evaluate the layers of a bit circuit on the CPU (SIMD & multi-threaded)

LogicCircuits.satisfies_flows_down_layers — Method

Pass flows down the layers of a bit circuit on the GPU

LogicCircuits.satisfies_flows_down_layers_cuda — Method

CUDA kernel for passing flows down circuit

LogicCircuits.satisfies_layers — Method

Evaluate the layers of a bit circuit on the CPU (SIMD & multi-threaded)

LogicCircuits.satisfies_layers — Method

Evaluate the layers of a bit circuit on the GPU

LogicCircuits.satisfies_layers_cuda — Method

CUDA kernel for circuit evaluation

LogicCircuits.save_bdd — Method

Saves a BDD as a file.

Supported file formats:

CNF (.cnf);

DNF (.dnf);

BDD (.bdd).

To save as any of these file formats, simply set the filename with the desired extension.

Keyword arguments are passed down to the open function.

LogicCircuits.save_bdd_ — Method

Save as BDD. Use the save_bdd function instead.

LogicCircuits.save_cnf — Method

Save as CNF. Use the save function instead.

LogicCircuits.save_dnf — Method

Save as DNF. Use the save function instead.

LogicCircuits.scope — Method

Returns all variables in this formula as a Vector{Int}.

LogicCircuits.scopeset — Method

Returns all variables in this formula as a Set{Int}.

LogicCircuits.sdd_mgr_for — Method

Obtain an SDD manager that can support compiling the given circuit

LogicCircuits.sdd_num_nodes — Method

Count the number of decision nodes in the SDD

LogicCircuits.sdd_size — Method

Count the number of elements in the decision nodes of the SDD

LogicCircuits.shallowhash — Function

Returns a shallow hash for the given node (not BDD as a whole).

LogicCircuits.shannon! — Method

Performs Shannon's Decomposition on the Bdd α, given a set of variables to isolate. Any decomposition that results in a ⊥ is discarded.

LogicCircuits.shannon — Method

Performs Shannon's Decomposition on the Bdd α, given a variable to isolate.

LogicCircuits.shannon — Method

Performs Shannon's Decomposition on the Bdd α, given a set of variables to isolate.

LogicCircuits.sibling — Method

Get the other child of a parent element

LogicCircuits.smooth — Method

smooth(root::Node)::Node

Create an equivalent smooth circuit from the given circuit.

LogicCircuits.smooth — Method

Smooth an sdd to a StructLogicCircuit

LogicCircuits.smooth — Method

smooth(root::StructLogicCircuit)::StructLogicCircuit

Create an equivalent smooth circuit from the given circuit.

LogicCircuits.smooth_node — Method

smooth_node(node::Node, missing_scope, lit_nodes)

Return a smooth version of the node where the missing_scope variables are added to the scope, using literals from lit_nodes

LogicCircuits.smooth_node — Method

smooth_node(node::StructLogicCircuit, parent_scope, scope, lit_nodes)

Return a smooth version of the node where the are added to the scope by filling the gap in vtrees, using literals from lit_nodes

LogicCircuits.split_candidates — Method

split_candidates(circuit::Node)::Tuple{Vector{Tuple{Node, Node}}, Dict{Node, BitSet}}

Return the edges and variables which can be splited on

LogicCircuits.split_step — Method

Split step

LogicCircuits.standardize_circuit — Method

Standardize the circuit:

Children of or nodes are and nodes.

Children of and nodes are or nodes.

Each and node has exactly two children.

Note: for circuits with parameters this function will not keep the parameters equavalent.

LogicCircuits.struct_learn — Method

Structure learning manager

LogicCircuits.sub — Method

Get the sub, that is, the second and last conjunct

LogicCircuits.terminal — Method

Returns a new terminal node of given boolean value.

LogicCircuits.to_int — Method

Returns 0 if x is not a literal; else returns an integer representation of x.

LogicCircuits.to_lit — Method

Returns α as an integer literal. Assumes α is a leaf node.

LogicCircuits.tree_formula_string — Method

Get the formula of a given circuit as a string, expanding the formula into a tree

LogicCircuits.unique⋁ — Method

Construct a unique decision gate for the given vtree

LogicCircuits.valuations — Method

Compute all possible valuations of scope V.

LogicCircuits.variable — Function

Get the variable in a vtree leaf

LogicCircuits.variable — Method

Get the logical variable in a given literal leaf node

LogicCircuits.varsubset — Method

Are the variables in n contained in the variables in m?

LogicCircuits.varsubset_left — Method

Are the variables in n contained in the left branch of m?

LogicCircuits.varsubset_right — Method

Are the variables in n contained in the right branch of m?

LogicCircuits.vtree — Method

Get the vtree corresponding to the argument

LogicCircuits.vtree_safe — Method

Get the vtree corresponding to the argument, or nothing if the node has no vtree

LogicCircuits.⋀_nodes — Method

Get the list of conjunction nodes in a given circuit

LogicCircuits.⋁_nodes — Method

Get the list of disjunction nodes in a given circuit