LogicCircuits

LogicCircuits.false_sdd — Constant

Canonical false Sdd node

LogicCircuits.structfalse — Constant

The unique splain tructured logical false constant

LogicCircuits.structtrue — Constant

The unique plain structured logical true constant

LogicCircuits.true_sdd — Constant

Canonical true Sdd node

LogicCircuits.ApplyArgs — Type

Representation of the arguments of an Apply call

LogicCircuits.ApplyCache — Type

Apply cache for the result of conjunctions and disjunctions

LogicCircuits.BitCircuit — Type

A bit circuit is a low-level representation of a logical circuit structure.

They are a "flat" representation of a circuit, essentially a bit string, that can be processed by lower level code (i.e., GPU kernels)

The wiring of the circuit is captured by two matrices: nodes and elements.

Nodes are either leafs or decision (disjunction) nodes in the circuit.

Elements are conjunction nodes in the circuit.

In addition, there is a vector of layers, where each layer is a list of node ids. Layer 1 is the leaf/input layer. Layer end is the circuit root.

And there is a vector of parents, pointing to element id parents of decision nodes.

Nodes are represented as a 4xN matrix where

nodes[1,:] is the first element id belonging to this decision

nodes[2,:] is the last element id belonging to this decision

nodes[3,:] is the first parent index belonging to this decision

nodes[4,:] is the last parent index belonging to this decision

Elements belonging to node i are elements[:, nodes[1,i]:nodes[2,i]] Parents belonging to node i are parents[nodes[3,i]:nodes[4,i]]

Elements are represented by a 3xE matrix, where

elements[1,:] is the decision node id (parents of the element),

elements[2,:] is the prime node id (child of the element)

elements[3,:] is the sub node id (child of the element)

LogicCircuits.BitCircuit — Method

construct a new BitCircuit accomodating the given number of features

LogicCircuits.ConstantGate — Type

A trait denoting constant leaf nodes of any type

LogicCircuits.Element — Type

Represents elements that are not yet compiled into conjunctions

LogicCircuits.GateType — Type

A trait hierarchy denoting types of nodes GateType defines an orthogonal type hierarchy of node types, not circuit types, so we can dispatch on node type regardless of circuit type. See @ref{https://docs.julialang.org/en/v1/manual/methods/#Trait-based-dispatch-1}

LogicCircuits.GateType — Method

Get the gate type trait of the given LogicCircuit

LogicCircuits.InnerGate — Type

A logical gate that is an inner node

LogicCircuits.LeafGate — Type

A logical gate that is a leaf node

LogicCircuits.LiteralGate — Type

A trait denoting literal leaf nodes of any type

LogicCircuits.LogicCircuit — Type

Root of the logic circuit node hierarchy

LogicCircuits.NodeId — Type

Integer identifier for a circuit node

LogicCircuits.NodeIds — Type

The BitCircuit ids associated with a node

LogicCircuits.PlainConstantNode — Type

A plain logical constant leaf node, representing true or false

LogicCircuits.PlainFalseNode — Type

Plain constant false node

LogicCircuits.PlainLiteralNode — Type

A plain logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.PlainLogicCircuit — Type

Root of the plain logic circuit node hierarchy

LogicCircuits.PlainLogicInnerNode — Type

A plain logical inner node

LogicCircuits.PlainLogicLeafNode — Type

A plain logical leaf node

LogicCircuits.PlainStructConstantNode — Type

A plain structured logical constant leaf node, representing true or false. These are the only structured nodes that don't have an associated vtree node (cf. SDD file format)

LogicCircuits.PlainStructFalseNode — Type

A plain structured logical false constant. Never construct one, use structfalse to access its unique instance

LogicCircuits.PlainStructLiteralNode — Type

A plain structured logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.PlainStructLogicCircuit — Type

Root of the plain structure logic circuit node hierarchy

LogicCircuits.PlainStructLogicInnerNode — Type

A plain structured logical inner node

LogicCircuits.PlainStructLogicLeafNode — Type

A plain structured logical leaf node

LogicCircuits.PlainStructTrueNode — Type

A plain structured logical true constant. Never construct one, use structtrue to access its unique instance

LogicCircuits.PlainStruct⋀Node — Type

A plain structured logical conjunction node

LogicCircuits.PlainStruct⋁Node — Type

A plain structured logical disjunction node

LogicCircuits.PlainTrueNode — Type

Plain constant true node

LogicCircuits.PlainVtree — Type

Root of the plain vtree node hierarchy

LogicCircuits.Plain⋀Node — Type

A plain logical conjunction node (And node)

LogicCircuits.Plain⋁Node — Type

A plain logical disjunction node (Or node)

LogicCircuits.Sdd — Type

Root of the trimmed Sdd circuit node hierarchy

LogicCircuits.SddConstantNode — Type

A SDD logical constant leaf node, representing true or false. These are the only structured nodes that don't have an associated vtree node (cf. SDD file format)

LogicCircuits.SddFalseNode — Type

A SDD logical false constant.

LogicCircuits.SddInnerNode — Type

A SDD logical inner node

LogicCircuits.SddLeafNode — Type

A SDD logical leaf node

LogicCircuits.SddLiteralNode — Type

A SDD logical literal leaf node, representing the positive or negative literal of its variable

LogicCircuits.SddMgr — Type

Root of the SDD manager node hierarchy

LogicCircuits.SddMgrInnerNode — Type

SDD manager inner vtree node for trimmed SDD nodes

LogicCircuits.SddMgrLeafNode — Type

SDD manager leaf vtree node for trimmed SDD nodes

LogicCircuits.SddTrueNode — Type

A SDD logical true constant.

LogicCircuits.Sdd⋀Node — Type

A SDD logical conjunction node

LogicCircuits.Sdd⋁Node — Type

A SDD logical disjunction node

LogicCircuits.StructLogicCircuit — Type

Root of the structure logic circuit node hierarchy

LogicCircuits.Unique⋁Cache — Type

Unique nodes cache for decision nodes

LogicCircuits.Vtree — Type

Root of the vtree node hiearchy

LogicCircuits.XYPartition — Type

Represent an XY-partition that has not yet been compiled into a disjunction

LogicCircuits.⋀Gate — Type

A trait denoting conjuction nodes of any type

LogicCircuits.⋁Gate — Type

A trait denoting disjunction nodes of any type

Base.deepcopy — Function

deepcopy(n::Node, depth::Int64)

Recursively create a copy circuit rooted at n to a certain depth depth

Base.merge — Method

merge(root::Node, or1::Node, or2::Node)

Merge two circuits.

Base.split — Method

split(root::Node, (or, and)::Tuple{Node, Node}, var::Var; depth=0, sanity_check=true)

Return the circuit after spliting on edge edge and variable var

LogicCircuits.Utils.children — Method

Get the children of a given inner node

LogicCircuits.Utils.depth — Method

Compute the path length from vtree node n to leaf node for variable var

LogicCircuits.Utils.find_inode — Method

Find an inner vtree node that has left in its left subtree and right in its right subtree. Supports nothing as a catch-all for any node

LogicCircuits.Utils.find_leaf — Method

Find the leaf in the vtree that represents the given variable

LogicCircuits.Utils.foldup — Method

foldup(node::LogicCircuit, 
    f_con::Function, 
    f_lit::Function, 
    f_a::Function, 
    f_o::Function, 
    ::Type{T}; nload = nload, nsave = nsave, reset=true)::T where {T}

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o through a callback from the children.

LogicCircuits.Utils.foldup_aggregate — Method

foldup_aggregate(node::LogicCircuit, 
    f_con::Function, 
    f_lit::Function, 
    f_a::Function, 
    f_o::Function, 
    ::Type{T};  nload = nload, nsave = nsave, reset=true)::T where T

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o in an aggregate vector from the children.

LogicCircuits.Utils.lca — Method

Compute the lowest common ancestor of two vtree nodes Warning: this method uses an incomplete varsubset check for descends_from and is only correct when v and w are part of the same larger vtree.

LogicCircuits.Utils.num_nodes — Method

How many nodes are indexed by a given bit circuit?

LogicCircuits.Utils.variables — Method

variables(root::LogicCircuit)::BitSet

Get a bitset of variables mentioned in the circuit root.

LogicCircuits.and_nodes — Method

Get the list of And nodes in a given circuit

LogicCircuits.balance_threads — Method

assign threads to examples and decision nodes so that everything is a power of 2

LogicCircuits.canonical_constants — Method

Construct a mapping from constants to their canonical node representation

LogicCircuits.canonical_literals — Method

Construct a mapping from literals to their canonical node representation

LogicCircuits.canonicalize — Method

Get the canonical compilation of the given XY Partition

LogicCircuits.canonicalize_compressed — Method

Get the canonical compilation of the given compressed XY Partition

LogicCircuits.clone — Method

Clone the or node and redirect one of its parents to the new copy

LogicCircuits.compile — Function

Create new circuit nodes in the given context.

LogicCircuits.compile — Method

Compile a given variable, literal, or constant

LogicCircuits.compress — Method

Compress a given XY Partition (merge elements with identical subs)

LogicCircuits.condition — Method

condition(root::Node, lit::Lit; callback::Function)::Node

Return the circuit conditioned on given literal constrains callback is called after modifying conjunction node

LogicCircuits.conjoin — Function

Conjoin nodes into a single circuit

LogicCircuits.conjoin — Method

Conjoin two SDDs

LogicCircuits.conjoin_cartesian — Method

Conjoin two SDDs when they respect the same vtree node

LogicCircuits.conjoin_descendent — Method

Conjoin two SDDs when one descends from the other

LogicCircuits.conjoin_indep — Method

Conjoin two SDDs in separate parts of the vtree

LogicCircuits.constant — Function

Get the logical constant in a given constant leaf node

LogicCircuits.constant — Method

Get the logical constant in a given constant leaf node

LogicCircuits.construct_new_sdd_literal — Method

Helper function to construct new SDD literal objects in SddMgrLeafNode constructor

LogicCircuits.disjoin — Function

Disjoin nodes into a single circuit

LogicCircuits.forget — Method

forget(root::Node, is_forgotten::Function)::Node

Forget variables from the circuit. Warning: this may or may not destroy the determinism property.

LogicCircuits.fully_factorized_circuit — Function

Generate a fully factorized circuit over the given range of variables

LogicCircuits.has_single_child — Method

Does the bitcircuit node have a single child?

LogicCircuits.infer_vtree — Method

infer_vtree(root::LogicCircuit)::Vtree

Infer circuits vtree if the circuit is struct-decomposable it. Otherwise return nothing.

LogicCircuits.init_satisfies — Method

Initialize values from the data (data frames)

LogicCircuits.iscanonical — Method

iscanonical(circuit::LogicCircuit, k::Int; verbose = false)

Does the given circuit have canonical Or gates, as determined by a probabilistic equivalence check?

LogicCircuits.isconstantgate — Method

Is the node a constant gate?

LogicCircuits.isdecomposable — Method

isdecomposable(root::LogicCircuit)::Bool

Is the circuit decomposable?

LogicCircuits.isdeterministic — Method

isdeterministic(root::LogicCircuit)::Bool

Is the circuit determinstic? Note: this function is generally intractable for large circuits.

LogicCircuits.isfalse — Method

Is the circuit syntactically equal to false?

LogicCircuits.isinnergate — Method

Is the node an inner gate?

LogicCircuits.isleafgate — Method

Is the node a leaf gate?

LogicCircuits.isliteralgate — Method

Is the node a literal gate?

LogicCircuits.ispositive — Method

Get the sign of the literal leaf node

LogicCircuits.issmooth — Method

issmooth(root::LogicCircuit)::Bool

Is the circuit smooth?

LogicCircuits.isstruct_decomposable — Method

isstruct_decomposable(root::LogicCircuit)::Bool

Is the circuit structured-decomposable?

LogicCircuits.istrue — Method

Is the circuit syntactically equal to true?

LogicCircuits.is⋀gate — Method

Is the node an And gate?

LogicCircuits.is⋁gate — Method

Is the node an Or gate?

LogicCircuits.literal — Method

Get the logical literal in a given literal leaf node

LogicCircuits.literal_nodes — Method

Get the list of literal nodes in a given circuit

LogicCircuits.mgr — Method

Get the manager of a Sdd node, which is its SddMgr vtree

LogicCircuits.model_count — Function

model_count(root::LogicCircuit, num_vars_in_scope::Int = num_variables(root))::BigInt

Get the model count of the circuit. The num_vars_in_scope is set to number of variables in the circuit, but sometimes need to set different values, for example, if not every variable is mentioned in the circuit.

LogicCircuits.negate — Method

Negate an SDD

LogicCircuits.num_elements — Method

Number of elements (conjunctions) in layer or bit circuit

LogicCircuits.or_nodes — Method

Get the list of or nodes in a given circuit

LogicCircuits.prime — Method

Get the prime, that is, the first conjunct

LogicCircuits.prob_equiv_signature — Method

prob_equiv_signature(circuit::LogicCircuit, k::Int)::Dict{Union{Var,Node},Signature}

Get a signature for each node using probabilistic equivalence checking. Note that this implentation may not have any formal guarantees as such.

LogicCircuits.propagate_constants — Method

propagate_constants(root::Node)

Remove all constant leafs from the circuit

LogicCircuits.random_split — Method

Randomly picking egde and variable from candidates

LogicCircuits.replace_node — Method

Replace node old with node new in circuit root

LogicCircuits.respects_vtree — Method

Does the circuit respect the given vtree? This function allows for constants in conjunctions, but only when a vtree node can be found where the left and right conjunct can be assigned to the left and right vtree.

LogicCircuits.sat_prob — Method

sat_prob(root::LogicCircuit; varprob::Function)::Rational{BigInt}

Get the probability that a random world satisties the circuit. Probability of each variable is given by varprob Function which defauls to 1/2 for every variable.

LogicCircuits.satisfies — Method

Evaluate satisfaction of the circuit bottom-up for a given input

LogicCircuits.satisfies_all — Function

Evaluate the circuit bottom-up for a given input and return the value of all nodes

LogicCircuits.satisfies_flows — Function

Compute the value and flow of each node

LogicCircuits.satisfies_flows_down — Function

When values of nodes have already been computed, do a downward pass computing the flows at each node

LogicCircuits.satisfies_flows_down_layers — Method

Evaluate the layers of a bit circuit on the CPU (SIMD & multi-threaded)

LogicCircuits.satisfies_flows_down_layers — Method

Pass flows down the layers of a bit circuit on the GPU

LogicCircuits.satisfies_flows_down_layers_cuda — Method

CUDA kernel for passing flows down circuit

LogicCircuits.satisfies_layers — Method

Evaluate the layers of a bit circuit on the CPU (SIMD & multi-threaded)

LogicCircuits.satisfies_layers — Method

Evaluate the layers of a bit circuit on the GPU

LogicCircuits.satisfies_layers_cuda — Method

CUDA kernel for circuit evaluation

LogicCircuits.sdd_mgr_for — Method

Obtain an SDD manager that can support compiling the given circuit

LogicCircuits.sdd_num_nodes — Method

Count the number of decision nodes in the SDD

LogicCircuits.sdd_size — Method

Count the number of elements in the decision nodes of the SDD

LogicCircuits.sibling — Method

Get the other child of a parent element

LogicCircuits.smooth — Method

smooth(root::Node)::Node

Create an equivalent smooth circuit from the given circuit.

LogicCircuits.smooth_node — Method

smooth_node(node::Node, missing_scope, lit_nodes)

Return a smooth version of the node where the missing_scope variables are added to the scope, using literals from lit_nodes

LogicCircuits.split_candidates — Method

split_candidates(circuit::Node)::Tuple{Vector{Tuple{Node, Node}}, Dict{Node, BitSet}}

Return the edges and variables which can be splited on

LogicCircuits.split_step — Method

Split step

LogicCircuits.standardize_circuit — Method

Standardize the circuit:

Children of or nodes are and nodes.

Children of and nodes are or nodes.

Each and node has exactly two children.

Note: for circuits with parameters this function will not keep the parameters equavalent.

LogicCircuits.struct_learn — Method

Structure learning manager

LogicCircuits.sub — Method

Get the sub, that is, the second and last conjunct

LogicCircuits.tree_formula_string — Method

Get the formula of a given circuit as a string, expanding the formula into a tree

LogicCircuits.unique⋁ — Method

Construct a unique decision gate for the given vtree

LogicCircuits.variable — Function

Get the variable in a vtree leaf

LogicCircuits.variable — Method

Get the logical variable in a given literal leaf node

LogicCircuits.variables_by_node — Method

Get the variable scope of each node in the circuit

LogicCircuits.varsubset — Method

Are the variables in n contained in the variables in m?

LogicCircuits.varsubset_left — Method

Are the variables in n contained in the left branch of m?

LogicCircuits.varsubset_right — Method

Are the variables in n contained in the right branch of m?

LogicCircuits.vtree — Method

Get the vtree corresponding to the argument

LogicCircuits.vtree_safe — Method

Get the vtree corresponding to the argument, or nothing if the node has no vtree

LogicCircuits.⋀_nodes — Method

Get the list of conjunction nodes in a given circuit

LogicCircuits.⋁_nodes — Method

Get the list of disjunction nodes in a given circuit