Utils

LogicCircuits.Utils — Module

Module with general utilities and missing standard library features that could be useful in any Julia project

LogicCircuits.Utils.AbstractBitVector — Type

Retro-fitted super type of all bit vectors

LogicCircuits.Utils.CuBitVector — Type

Custom CUDA version of BitVector (lacking lots of functionality, just a container for now).

LogicCircuits.Utils.Dag — Type

A node in a directed acyclic graph (of which it is the root)

LogicCircuits.Utils.Inner — Type

The trait of inner nodes (nodes that have children)

LogicCircuits.Utils.Leaf — Type

The trait of leaf nodes (nodes without children)

LogicCircuits.Utils.Lit — Type

Literals are represented as 32-bit signed integers. Positive literals are positive integers identical to their variable. Negative literals are their negations. Integer 0 should not be used to represent literals.

LogicCircuits.Utils.Node — Type

A node in a directed graph. This type hierarchy is organized according to the type of graph the node is part of.

LogicCircuits.Utils.NodeType — Type

A trait hierarchy denoting types of Dag nodes NodeType defines an orthogonal type hierarchy of node types, so we can dispatch on node type regardless of the graph type. See @ref{https://docs.julialang.org/en/v1/manual/methods/#Trait-based-dispatch-1}

LogicCircuits.Utils.NodeType — Method

Get the node type trait of the given Node

LogicCircuits.Utils.Tree — Type

A node in a tree (of which it is the root)

LogicCircuits.Utils.Var — Type

Variables are represented as 32-bit unsigned integers

Base.eltype — Method

Find a type that can capture all feature values

Base.filter — Method

Retrieve list of nodes in graph matching predicate p

Base.foreach — Method

Apply a function to each node in a graph, bottom up

Base.in — Method

Is the node contained in the Dag?

Base.isdisjoint — Method

Are the given sets disjoint (no shared elements)?

Base.parent — Method

Get the parent of a given tree node (or nothing if the node is root)

DataFrames.groupby — Method

Group the elements of list by their values according to function f

LogicCircuits.Utils.:¬ — Function

Logical negation

LogicCircuits.Utils.:⇐ — Method

Material logical implication (reverse)

LogicCircuits.Utils.:⇒ — Method

Material logical implication

LogicCircuits.Utils.:⇔ — Method

Bidirectional logical implication

LogicCircuits.Utils.:∧ — Function

Logical conjunction

LogicCircuits.Utils.:∨ — Function

Logical disjunction

LogicCircuits.Utils.add_sample_weights — Method

Add sample weights by adding a named column weight at the end of the DataFrame.

LogicCircuits.Utils.always — Method

An array of 100% probabilities for the given element type

LogicCircuits.Utils.bagging_dataset — Method

Dataset bagging

LogicCircuits.Utils.batch — Function

Create mini-batches

LogicCircuits.Utils.batch_size — Method

Batch size of the dataset

LogicCircuits.Utils.bits_per_pixel — Method

Normalize the given log-likelihood as bits per pixel in data

LogicCircuits.Utils.children — Function

Get the children of a given inner node

LogicCircuits.Utils.chunks — Method

Retrieve chunks of bit vector

LogicCircuits.Utils.clear_data — Method

Set all the data fields in a circuit to nothing

LogicCircuits.Utils.count_parents — Method

Set the counter field of each node to its number of parents

LogicCircuits.Utils.depth — Method

Compute the length of the path from a tree node to the leaf following the branching function

LogicCircuits.Utils.example — Method

Get the ith example

LogicCircuits.Utils.feature_values — Method

Get the ith feature values

LogicCircuits.Utils.find_inode — Method

Find a binary inner node that has left in its left subtree and right in its right subtree. Supports nothing as a catch-all for either left or right. Returns nothing if no such node exists.

LogicCircuits.Utils.find_leaf — Method

Find the leaf in the tree by follwing the branching function

LogicCircuits.Utils.foldup — Method

foldup(node::Dag, 
    f_leaf::Function, 
    f_inner::Function, 
    ::Type{T}; nload = nload, nsave = nsave, reset=true)::T where {T}

Compute a function bottom-up on the graph. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner as a function on the children.

LogicCircuits.Utils.foldup_rec — Method

Compute a function bottom-up on the graph, without resetting the counter. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner as a function on the children.

LogicCircuits.Utils.foreach_down — Method

Apply a function to each node in a graph, top down

LogicCircuits.Utils.foreach_down_rec — Method

Apply a function to each node in a graph, top down, without setting the counter first

LogicCircuits.Utils.foreach_rec — Method

Apply a function to each node in a graph, bottom up, without resetting the counter

LogicCircuits.Utils.foreach_reset — Method

Apply a function to each node in a graph, bottom up, while resetting the counter

LogicCircuits.Utils.fully_factorized_log_likelihood — Method

Computer the per-example log-likelihood of a fully factorized ML model on Bool data

LogicCircuits.Utils.get_bit — Method

Retrieve the jth bit from a BitVector chunk

LogicCircuits.Utils.get_weights — Method

Get the weights from a weighted dataset.

LogicCircuits.Utils.has_children — Method

Does the node have children?

LogicCircuits.Utils.has_parent — Method

Does the node have a parent?

LogicCircuits.Utils.impute — Method

Return a copy of Imputed values of X (potentially statistics from another DataFrame)

For example, to impute using same DataFrame:

impute(X; method=:median)

If you want to use another DataFrame to provide imputation statistics:

impute(test_x, train_x; method=:mean)

Supported methods are :median, :mean, :one, :zero

LogicCircuits.Utils.init_array — Method

An array of undetermined values (fast) for the given element type

LogicCircuits.Utils.innernodes — Method

Get the list of inner nodes in a given graph

LogicCircuits.Utils.inode_stats — Method

Give count of types and fan-ins of inner nodes in the graph

LogicCircuits.Utils.inodes — Method

Get the list of inner nodes in a given graph

LogicCircuits.Utils.isbatched — Method

Is the dataset batched?

LogicCircuits.Utils.isbinarydata — Method

Is the dataset binary?

LogicCircuits.Utils.iscomplete — Method

Is the data complete (no missing values)?

LogicCircuits.Utils.isequal_local — Method

Is one node equal to another locally, ignoring children?

LogicCircuits.Utils.isfpdata — Method

Is the dataset consisting of floating point data?

LogicCircuits.Utils.isgpu — Method

Check whether data resides on the GPU

LogicCircuits.Utils.isinner — Method

Is the node an Inner node?

LogicCircuits.Utils.isleaf — Method

Is the node a leaf node?

LogicCircuits.Utils.isroot — Method

Is the node the root of its tree?

LogicCircuits.Utils.issingle — Method

Is the argument of length 1; a singleton?

LogicCircuits.Utils.issomething — Method

Is the argument not nothing?

LogicCircuits.Utils.isweighted — Method

Is the dataset weighted?

LogicCircuits.Utils.lca — Method

Find the least common ancestor. Assumes the Tree has access to a parent. A given descends_from function is required to quickly check whether a node is an ancestor.

LogicCircuits.Utils.leaf_stats — Method

Give count of types of leaf nodes in the graph

LogicCircuits.Utils.leafnodes — Method

Get the list of leaf nodes in a given graph

LogicCircuits.Utils.left_most_descendent — Method

Return the left-most descendent.

LogicCircuits.Utils.linearize — Method

Order the Dag's nodes bottom-up in a list (with optional element type)

LogicCircuits.Utils.lit2var — Method

Convert a literal its variable, removing the sign of the literal

LogicCircuits.Utils.ll_per_example — Method

Normalize the given log-likelihood by the number of examples in data

LogicCircuits.Utils.make_missing_mcar — Method

make_missing_mcar(d::DataFrame; keep_prob::Float64=0.8)

Returns a copy of dataframe with making some features missing as MCAR, with keep_prob as probability of keeping each feature.

LogicCircuits.Utils.map_values — Method

Map the values in the dictionary, retaining the same keys

LogicCircuits.Utils.marginal_prob — Method

Compute the marginal prob of each feature in a binary dataset.

LogicCircuits.Utils.never — Method

An array of 0% probabilities for the given element type

LogicCircuits.Utils.nload — Method

Default getter to obtain data associated with a node

LogicCircuits.Utils.node_stats — Method

Give count of types and fan-ins of all nodes in the graph

LogicCircuits.Utils.noop — Method

Function that does nothing

LogicCircuits.Utils.nsave — Method

Default setter to assign data associated with a node

LogicCircuits.Utils.num_children — Method

Get the number of children of a given inner node

LogicCircuits.Utils.num_chunks — Method

For binary complete data, how many UInt64 bit strings are needed to store one feature?

LogicCircuits.Utils.num_edges — Method

Number of edges in the Dag

LogicCircuits.Utils.num_examples — Method

num_examples(df::DataFrame)

Number of examples in data

LogicCircuits.Utils.num_features — Method

num_features(df::DataFrame)

Number of features in the data

LogicCircuits.Utils.num_nodes — Method

num_nodes(node::Dag)

Count the number of nodes in the Dag

LogicCircuits.Utils.num_variables — Method

Number of variables in the data structure

LogicCircuits.Utils.order_asc — Method

Order the arguments in a tuple in ascending order

LogicCircuits.Utils.print_tree — Function

Print the given tree

LogicCircuits.Utils.pushrand! — Method

Push element into random position in vectorv

LogicCircuits.Utils.reset_counter — Function

Set the counter field throughout this graph (assumes no node in the graph already has the value)

LogicCircuits.Utils.right_most_descendent — Method

Return the right-most descendent.

LogicCircuits.Utils.root — Method

Get the root of the given tree node

LogicCircuits.Utils.same_device — Method

Ensure that x resides on the same device as data

LogicCircuits.Utils.shuffle_examples — Method

shuffle_examples(df::DataFrame)

Shuffle the examples in the data

LogicCircuits.Utils.similar! — Method

Reuse a given array if it has the right type and size, otherwise make a new one

LogicCircuits.Utils.soften — Function

Turn binary data into floating point data close to 0 and 1.

LogicCircuits.Utils.split_sample_weights — Method

Split a weighted dataset into unweighted dataset and its corresponding weights.

LogicCircuits.Utils.subseteq_fast — Method

Replacement for BitSet.⊆ that does not allocate a new BitSet

LogicCircuits.Utils.threshold — Method

Threshold a numeric dataset making it binary

LogicCircuits.Utils.to_cpu — Method

Move data to the CPU

LogicCircuits.Utils.to_gpu — Method

Move data to the GPU

LogicCircuits.Utils.tree_num_edges — Method

tree_num_edges(node::Dag)::BigInt

Compute the number of edges in the tree-unfolding of the Dag.

LogicCircuits.Utils.tree_num_nodes — Method

tree_num_nodes(node::Dag)::BigInt

Compute the number of nodes in of a tree-unfolding of the Dag.

LogicCircuits.Utils.uniform — Method

An array of uniform probabilities

LogicCircuits.Utils.var2lit — Method

Convert a variable to the corresponding positive literal

LogicCircuits.Utils.variables — Function

Get the BitSet of variables in the data structure

StatsFuns.logsumexp — Method

Marginalize out dimensions dims from log-probability tensor