Utils

LogicCircuits.Utils — Module

Module with general utilities and missing standard library features that could be useful in any Julia project

Base.typejoin — Method

Get the most specific type parameter possible for an array

LogicCircuits.Utils.children — Method

Get the children of a given inner node

LogicCircuits.Utils.copy_with_eltype — Method

Copy the array while changing the element type

LogicCircuits.Utils.flip_bit — Method

Flip the bit field throughout this circuit (or ensure it is set to given value)

LogicCircuits.Utils.foldup — Method

Compute a function bottom-up on the circuit. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner as a function on the children.

LogicCircuits.Utils.foldup_aggregate — Method

Compute a function bottom-up on the circuit. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner in aggregate as a vector from the children.

LogicCircuits.Utils.grapheltype — Method

Get the type of node contained in this graph

LogicCircuits.Utils.has_children — Method

Does the node have children?

LogicCircuits.Utils.innernodes — Method

Get the list of inner nodes in a given graph

LogicCircuits.Utils.inode_stats — Method

Give count of types and fan-ins of inner nodes in the graph

LogicCircuits.Utils.inodes — Method

Get the list of inner nodes in a given graph

LogicCircuits.Utils.isequal_local — Function

Is one node equal to another locally, ignoring children?

LogicCircuits.Utils.isequal_unordered — Method

Is one unordered tree equal to another?

LogicCircuits.Utils.isinner — Method

Is the node an Inner node?

LogicCircuits.Utils.isleaf — Method

Is the node a leaf node?

LogicCircuits.Utils.issomething — Method

Is the argument not nothing?

LogicCircuits.Utils.lca — Method

Find the least common ancestor (assumes the graph has a parent pointer and a list of descendents)

LogicCircuits.Utils.leaf_stats — Method

Give count of types of leaf nodes in the graph

LogicCircuits.Utils.leafnodes — Method

Get the list of leaf nodes in a given graph

LogicCircuits.Utils.left_most_child — Method

Return the leftmost child.

LogicCircuits.Utils.lower_element_type — Method

Specialize the type parameter of an array to be most specific

LogicCircuits.Utils.node2dag — Method

Rebuild a DAG's linear bottom-up order from a new root node

LogicCircuits.Utils.node_stats — Method

Give count of types and fan-ins of all nodes in the graph

LogicCircuits.Utils.num_children — Method

Get the number of children of a given inner node

LogicCircuits.Utils.num_edges — Method

Number of edges in the graph

LogicCircuits.Utils.num_nodes — Method

Number of nodes in the graph

LogicCircuits.Utils.pushrand! — Method

Push element into random position

LogicCircuits.Utils.right_most_child — Method

Return the rightmost child.

LogicCircuits.Utils.tree_num_nodes — Method

Compute the number of nodes in of a tree-unfolding of the DAG.

StatsFuns.logsumexp — Method

Marginalize out dimensions dims from log-probability tensor

LogicCircuits.Utils.Downpass — Type

Compute a function top down on the circuit. f_root is called on the root node, f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed down the circuit and given to f_inner as a value from the parent.

Base.filter — Method

Retrieve list of nodes in circuit matching predicate p

Base.foreach — Method

Apply a function to each node in a circuit, bottom up

Base.isequal — Method

Is one ordered tree equal to another?

LogicCircuits.Utils.expand_product — Method

generate an expression for the loop-unrolled product of x1 and xs...

LogicCircuits.Utils.expand_product — Method

generate an expression for the loop-unrolled product of xs...