Public API

Documentation for LogicCircuits.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Public API

Contents

Index

Public Interface

Index

LogicCircuits.LoadSave.load_logic_circuit

LogicCircuits.LoadSave.load_smooth_logic_circuit

Public Interface

LogicCircuits.LoadSave.load_logic_circuit — Function

Load a logical circuit from file. Support file formats:

".sdd" for SDD files

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.LoadSave.load_smooth_logic_circuit — Function

Load a smooth logical circuit from file. Support file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

Note

This list is not complete yet, will be adding more here as the public API gets finalized.