Public APIs

This page lists documentation for the most commonly used public APIs of LogicCircuits.jl. Visit the internals section for a auto generated documentation for more public API and internal APIs.

Public APIs

Loading Circuits

Saving Circuits

Circuit Properties

Circuit Queries

Circuit Operations

Circuit Transformations

Compilation

Loading Circuits

LogicCircuits.LoadSave.load_cnf — Function

load_cnf(file::String; dual=false)::PlainLogicCircuit

Load a CNF/DNF as a logic circuit from file. Suppported file formats:

".cnf" for CNF files (dual=False)

".dnf" for DNF files (dual=True)

LogicCircuits.LoadSave.load_dnf — Function

load_dnf(file::String)::PlainLogicCircuit

Load a DNF as a logical circuit from file.

LogicCircuits.LoadSave.load_logic_circuit — Function

load_logic_circuit(file::String)::PlainLogicCircuit

Load a logical circuit from file. Support file formats:

".sdd" for SDD files

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.LoadSave.load_smooth_logic_circuit — Function

load_smooth_logic_circuit(file::String)::PlainLogicCircuit

Load a smooth logic circuit from file. Support file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

LogicCircuits.LoadSave.load_struct_smooth_logic_circuit — Function

load_struct_smooth_logic_circuit(circuit_file::String, vtree_file::String)::Tuple{StructLogicCircuit,PlainVtree}

Load a smooth structured logic circuit and its vtree from file. Supported circuit file formats:

".psdd" for PSDD files

".circuit" for Logistic Circuit files

Supported vtree file formats:

".vtree" for Vtree files

Saving Circuits

LogicCircuits.LoadSave.save_circuit — Function

save_circuit(name::String, circuit, vtree)

Save a circuit to file. Currently defaults to and only supports saving '.sdd' format.

LogicCircuits.LoadSave.save_vtree — Function

save_vtree(vtree::PlainVtree, file::AbstractString)

Saves a vtree in the given file path based on file format. Supported formats:

".vtree" for Vtree files

".dot" for dot files

LogicCircuits.LoadSave.save_as_sdd — Function

save_as_sdd(name::String, circuit, vtree)

Save a SDD circuit to file. File name should end with '.sdd'.

LogicCircuits.LoadSave.save_as_dot — Function

save_as_dot(circuit::LogicCircuit, file::String)

Save logic circuit in 'dot' file format.

LogicCircuits.LoadSave.save_as_tex — Function

save_as_tex(circuit::LogicCircuit, file::String; **other_args)

Save logic circuit as a LaTeX TikZ .tex file. Argument V is a label map for each vertex, E is a label map for each edge, ⋀_style is the TikZ conjunction node style to use, ⋁_style is the TikZ disjunction node style to use, node_ysep is the y-axis separation distance for nodes, node_xsep for the x-axis, edge_turn is how far down should edges "bend", verbose if set to true prints comments to the LaTeX file for better readability.

Note: save_as_tex does not work well on large circuits, and the resulting LaTeX file should serve as a first sketch in (serious) need for adjustments.

LogicCircuits.LoadSave.save_as_dot2tex — Function

save_as_dot2tex(circuit::LogicCircuit, file::String; **other_args)

Save logic circuit as a LaTeX TikZ .tex file using the dot2tex engine.

Argument V is a label map for each vertex; E is a label map for each edge; ⋀_style, ⋁_style, lit_style, ⊤_style, and ⊥_style are styles for each node type; edge_turn is how far down should edges "bend"; node_xsep and node_ysep are x and y-axis separation sizes for terminal nodes.

Note: save_as_tex does not work well on large circuits, and the resulting LaTeX file should serve as a first sketch in (serious) need for adjustments.

Circuit Properties

LogicCircuits.issmooth — Function

issmooth(root::LogicCircuit)::Bool

Is the circuit smooth?

LogicCircuits.isdecomposable — Function

isdecomposable(root::LogicCircuit)::Bool

Is the circuit decomposable?

LogicCircuits.isdeterministic — Function

isdeterministic(root::LogicCircuit)::Bool

Is the circuit determinstic? Note: this function is generally intractable for large circuits.

LogicCircuits.iscanonical — Function

iscanonical(circuit::LogicCircuit, k::Int; verbose = false)

Does the given circuit have canonical Or gates, as determined by a probabilistic equivalence check?

Circuit Queries

LogicCircuits.sat_prob — Function

sat_prob(root::LogicCircuit; varprob::Function)::Rational{BigInt}

Get the probability that a random world satisties the circuit. Probability of each variable is given by varprob Function which defauls to 1/2 for every variable.

LogicCircuits.model_count — Function

model_count(root::LogicCircuit, num_vars_in_scope::Int = num_variables(root))::BigInt

Get the model count of the circuit. The num_vars_in_scope is set to number of variables in the circuit, but sometimes need to set different values, for example, if not every variable is mentioned in the circuit.

Circuit Operations

LogicCircuits.Utils.foldup — Function

foldup(node::LogicCircuit, 
    f_con::Function, 
    f_lit::Function, 
    f_a::Function, 
    f_o::Function, 
    ::Type{T}; nload = nload, nsave = nsave, reset=true)::T where {T}

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o through a callback from the children.

foldup(node::Dag, 
    f_leaf::Function, 
    f_inner::Function, 
    ::Type{T}; nload = nload, nsave = nsave, reset=true)::T where {T}

Compute a function bottom-up on the graph. f_leaf is called on leaf nodes, and f_inner is called on inner nodes. Values of type T are passed up the circuit and given to f_inner as a function on the children.

LogicCircuits.Utils.foldup_aggregate — Function

foldup_aggregate(node::LogicCircuit, 
    f_con::Function, 
    f_lit::Function, 
    f_a::Function, 
    f_o::Function, 
    ::Type{T};  nload = nload, nsave = nsave, reset=true)::T where T

Compute a function bottom-up on the circuit. f_con is called on constant gates, f_lit is called on literal gates, f_a is called on conjunctions, and f_o is called on disjunctions. Values of type T are passed up the circuit and given to f_a and f_o in an aggregate vector from the children.

Circuit Transformations

LogicCircuits.smooth — Function

smooth(root::Node)::Node

Create an equivalent smooth circuit from the given circuit.

LogicCircuits.condition — Function

condition(root::Node, lit::Lit; callback::Function)::Node

Return the circuit conditioned on given literal constrains callback is called after modifying conjunction node

LogicCircuits.forget — Function

forget(root::Node, is_forgotten::Function)::Node

Forget variables from the circuit. Warning: this may or may not destroy the determinism property.

Base.deepcopy — Function

deepcopy(n::Node, depth::Int64)

Recursively create a copy circuit rooted at n to a certain depth depth

Compilation